sqlparser::ast

Enum ColumnOption

pub enum ColumnOption {

    Null,
    NotNull,
    Default(Expr),
    Materialized(Expr),
    Ephemeral(Option<Expr>),
    Alias(Expr),
    Unique {
        is_primary: bool,
        characteristics: Option<ConstraintCharacteristics>,
    },
    ForeignKey {
        foreign_table: ObjectName,
        referred_columns: Vec<Ident>,
        on_delete: Option<ReferentialAction>,
        on_update: Option<ReferentialAction>,
        characteristics: Option<ConstraintCharacteristics>,
    },
    Check(Expr),
    DialectSpecific(Vec<Token>),
    CharacterSet(ObjectName),
    Collation(ObjectName),
    Comment(String),
    OnUpdate(Expr),
    Generated {
        generated_as: GeneratedAs,
        sequence_options: Option<Vec<SequenceOptions>>,
        generation_expr: Option<Expr>,
        generation_expr_mode: Option<GeneratedExpressionMode>,
        generated_keyword: bool,
    },
    Options(Vec<SqlOption>),
    Identity(IdentityPropertyKind),
    OnConflict(Keyword),
    Policy(ColumnPolicy),
    Tags(TagsColumnOption),
    Srid(Box<Expr>),

}

ColumnOptions are modifiers that follow a column definition in a CREATE TABLE statement.

Variants

Null

NULL

NotNull

NOT NULL

Default(Expr)

DEFAULT <restricted-expr>

Materialized(Expr)

MATERIALIZE <expr> Syntax: b INT MATERIALIZE (a + 1)

ClickHouse

Ephemeral(Option<Expr>)

EPHEMERAL [<expr>]

ClickHouse

Alias(Expr)

ALIAS <expr>

ClickHouse

Unique

{ PRIMARY KEY | UNIQUE } [<constraint_characteristics>]

Fields

is_primary: bool

characteristics: Option<ConstraintCharacteristics>

ForeignKey

A referential integrity constraint ([FOREIGN KEY REFERENCES <foreign_table> (<referred_columns>) { [ON DELETE <referential_action>] [ON UPDATE <referential_action>] | [ON UPDATE <referential_action>] [ON DELETE <referential_action>] } [<constraint_characteristics>] ).

Fields

foreign_table: ObjectName

referred_columns: Vec<Ident>

on_delete: Option<ReferentialAction>

on_update: Option<ReferentialAction>

characteristics: Option<ConstraintCharacteristics>

Check(Expr)

CHECK (<expr>)

DialectSpecific(Vec<Token>)

Dialect-specific options, such as:

MySQL’s AUTO_INCREMENT or SQLite’s AUTOINCREMENT

…

CharacterSet(ObjectName)

Collation(ObjectName)

Comment(String)

OnUpdate(Expr)

Generated

Generateds are modifiers that follow a column definition in a CREATE TABLE statement.

Fields

generated_as: GeneratedAs

sequence_options: Option<Vec<SequenceOptions>>

generation_expr: Option<Expr>

generation_expr_mode: Option<GeneratedExpressionMode>

generated_keyword: bool

false if ‘GENERATED ALWAYS’ is skipped (option starts with AS)

Options(Vec<SqlOption>)

BigQuery specific: Explicit column options in a view 1 or table 2 Syntax

OPTIONS(description="field desc")

Identity(IdentityPropertyKind)

Creates an identity or an autoincrement column in a table. Syntax

{ IDENTITY | AUTOINCREMENT } [ (seed , increment) | START num INCREMENT num ] [ ORDER | NOORDER ]

OnConflict(Keyword)

SQLite specific: ON CONFLICT option on column definition https://www.sqlite.org/lang_conflict.html

Policy(ColumnPolicy)

Snowflake specific: an option of specifying security masking or projection policy to set on a column. Syntax:

[ WITH ] MASKING POLICY <policy_name> [ USING ( <col_name> , <cond_col1> , ... ) ]
[ WITH ] PROJECTION POLICY <policy_name>

Tags(TagsColumnOption)

Snowflake specific: Specifies the tag name and the tag string value. Syntax:

[ WITH ] TAG ( <tag_name> = '<tag_value>' [ , <tag_name> = '<tag_value>' , ... ] )

Srid(Box<Expr>)

MySQL specific: Spatial reference identifier Syntax:

CREATE TABLE geom (g GEOMETRY NOT NULL SRID 4326);