sqlparser::ast

Enum Set

pub enum Set {
    SingleAssignment {
        scope: Option<ContextModifier>,
        hivevar: bool,
        variable: ObjectName,
        values: Vec<Expr>,
    },
    ParenthesizedAssignments {
        variables: Vec<ObjectName>,
        values: Vec<Expr>,
    },
    MultipleAssignments {
        assignments: Vec<SetAssignment>,
    },
    SetSessionParam(SetSessionParamKind),
    SetRole {
        context_modifier: Option<ContextModifier>,
        role_name: Option<Ident>,
    },
    SetTimeZone {
        local: bool,
        value: Expr,
    },
    SetNames {
        charset_name: Ident,
        collation_name: Option<String>,
    },
    SetNamesDefault {},
    SetTransaction {
        modes: Vec<TransactionMode>,
        snapshot: Option<Value>,
        session: bool,
    },
}

Variants

SingleAssignment

SQL Standard-style SET a = 1;

Fields

scope: Option<ContextModifier>

hivevar: bool

variable: ObjectName

values: Vec<Expr>

ParenthesizedAssignments

Snowflake-style SET (a, b, ..) = (1, 2, ..);

Fields

variables: Vec<ObjectName>

values: Vec<Expr>

MultipleAssignments

MySQL-style SET a = 1, b = 2, ..;

Fields

assignments: Vec<SetAssignment>

SetSessionParam(SetSessionParamKind)

MS-SQL session

See https://learn.microsoft.com/en-us/sql/t-sql/statements/set-statements-transact-sql

SetRole

SET [ SESSION | LOCAL ] ROLE role_name

Sets session state. Examples: ANSI, Postgresql, MySQL, and Oracle

Fields

context_modifier: Option<ContextModifier>

Non-ANSI optional identifier to inform if the role is defined inside the current session (SESSION) or transaction (LOCAL).

role_name: Option<Ident>

Role name. If NONE is specified, then the current role name is removed.

SetTimeZone

SET TIME ZONE <value>

Note: this is a PostgreSQL-specific statements SET TIME ZONE <value> is an alias for SET timezone TO <value> in PostgreSQL However, we allow it for all dialects.

Fields

local: bool

value: Expr

SetNames

SET NAMES 'charset_name' [COLLATE 'collation_name']

Fields

charset_name: Ident

collation_name: Option<String>

SetNamesDefault

SET NAMES DEFAULT

Note: this is a MySQL-specific statement.

SetTransaction

SET TRANSACTION ...

Fields

modes: Vec<TransactionMode>

snapshot: Option<Value>

session: bool