sqlparser::ast

Module table_constraints

SQL Abstract Syntax Tree (AST) types for table constraints

Structs

CheckConstraint

A CHECK constraint ([ CONSTRAINT <name> ] CHECK (<expr>) [[NOT] ENFORCED]).

ForeignKeyConstraint

A referential integrity constraint ([ CONSTRAINT <name> ] FOREIGN KEY (<columns>) REFERENCES <foreign_table> (<referred_columns>) [ MATCH { FULL | PARTIAL | SIMPLE } ] { [ON DELETE <referential_action>] [ON UPDATE <referential_action>] | [ON UPDATE <referential_action>] [ON DELETE <referential_action>] }).

FullTextOrSpatialConstraint

MySQLs fulltext definition. Since the SPATIAL definition is exactly the same, and MySQL displays both the same way, it is part of this definition as well.

IndexConstraint

MySQLs index definition for index creation. Not present on ANSI so, for now, the usage is restricted to MySQL, as no other dialects that support this syntax were found.

PrimaryKeyConstraint

MySQL definition for PRIMARY KEY constraints statements:

UniqueConstraint

Unique constraint definition.

Enums

TableConstraint

A table-level constraint, specified in a CREATE TABLE or an ALTER TABLE ADD <constraint> statement.