sqlparser::ast

Trait Spanned

pub trait Spanned {
    // Required method
    fn span(&self) -> Span;
}

Trait for AST nodes that have a source location information.

Notes:

Source Span are not yet complete. They may be missing:

keywords or other tokens

span information entirely, in which case they return Span::empty().

Note Some impl blocks (rendered below) are annotated with which nodes are missing spans. See this ticket for additional information and status.

Example

let dialect = GenericDialect {};
let sql = r#"SELECT *
  FROM table_1"#;
let statements = Parser::new(&dialect)
  .try_with_sql(sql)?
  .parse_statements()?;
// Get the span of the first statement (SELECT)
let span = statements[0].span();
// statement starts at line 1, column 1 (1 based, not 0 based)
assert_eq!(span.start, Location::new(1, 1));
// statement ends on line 2, column 15
assert_eq!(span.end, Location::new(2, 15));

Required Methods

fn span(&self) -> Span

Return the Span (the minimum and maximum Location) for this AST node, by recursively combining the spans of its children.