Trait sqlparser::ast::Visit

pub trait Visit {
    // Required method
    fn visit<V: Visitor>(&self, visitor: &mut V) -> ControlFlow<V::Break>;
}

A type that can be visited by a Visitor. See Visitor for recursively visiting parsed SQL statements.

Note

This trait should be automatically derived for sqlparser AST nodes using the Visit proc macro.

#[cfg_attr(feature = "visitor", derive(Visit, VisitMut))]

Required Methods

fn visit<V: Visitor>(&self, visitor: &mut V) -> ControlFlow<V::Break>