sqlparser::ast

Trait VisitMut

pub trait VisitMut {
    // Required method
    fn visit<V: VisitorMut>(&mut self, visitor: &mut V) -> ControlFlow<V::Break>;
}

A type that can be visited by a VisitorMut. See VisitorMut for recursively visiting parsed SQL statements.

Note

This trait should be automatically derived for sqlparser AST nodes using the VisitMut proc macro.

#[cfg_attr(feature = "visitor", derive(Visit, VisitMut))]

Required Methods

fn visit<V: VisitorMut>(&mut self, visitor: &mut V) -> ControlFlow<V::Break>

Mutably visit this node with the provided VisitorMut.

Implementations should call the appropriate mutable visitor hooks to traverse and allow in-place mutation of child nodes. Returning a ControlFlow value permits early termination of the traversal.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".