sqlparser

Module dialect

Re-exports

pub use crate::keywords;

Structs

AnsiDialect

A Dialect for ANSI SQL.

BigQueryDialect

A Dialect for Google Bigquery

ClickHouseDialect

DatabricksDialect

A Dialect for Databricks SQL

DuckDbDialect

A Dialect for DuckDB

GenericDialect

A permissive, general purpose Dialect, which parses a wide variety of SQL statements, from many different dialects.

HiveDialect

A Dialect for Hive.

MsSqlDialect

A Dialect for Microsoft SQL Server

MySqlDialect

A Dialect for MySQL

PostgreSqlDialect

A Dialect for PostgreSQL

RedshiftSqlDialect

A Dialect for RedShift

SQLiteDialect

A Dialect for SQLite

SnowflakeDialect

A Dialect for Snowflake

Enums

Precedence

This represents the operators for which precedence must be defined

Traits

Dialect

Encapsulates the differences between SQL implementations.

Functions

dialect_from_str

Returns the built in Dialect corresponding to dialect_name.