sqlparser

Module keywords

This module defines:

a list of constants for every keyword

an ALL_KEYWORDS array with every keyword in it This is not a list of reserved keywords: some of these can be parsed as identifiers if the parser decides so. This means that new keywords can be added here without affecting the parse result.

As a matter of fact, most of these keywords are not used at all and could be removed.

a RESERVED_FOR_TABLE_ALIAS array with keywords reserved in a “table alias” context.

Enums

Keyword

An enumeration of SQL keywords recognized by the parser.

Constants

ABORT

The ABORT SQL keyword.

ABS

The ABS SQL keyword.

ABSENT

The ABSENT SQL keyword.

ABSOLUTE

The ABSOLUTE SQL keyword.

ACCEPTANYDATE

The ACCEPTANYDATE SQL keyword.

ACCEPTINVCHARS

The ACCEPTINVCHARS SQL keyword.

ACCESS

The ACCESS SQL keyword.

ACCOUNT

The ACCOUNT SQL keyword.

ACTION

The ACTION SQL keyword.

ADD

The ADD SQL keyword.

ADDQUOTES

The ADDQUOTES SQL keyword.

ADMIN

The ADMIN SQL keyword.

AFTER

The AFTER SQL keyword.

AGAINST

The AGAINST SQL keyword.

AGGREGATE

The AGGREGATE SQL keyword.

AGGREGATION

The AGGREGATION SQL keyword.

ALERT

The ALERT SQL keyword.

ALGORITHM

The ALGORITHM SQL keyword.

ALIAS

The ALIAS SQL keyword.

ALIGNMENT

The ALIGNMENT SQL keyword.

ALL

The ALL SQL keyword.

ALLOCATE

The ALLOCATE SQL keyword.

ALLOWOVERWRITE

The ALLOWOVERWRITE SQL keyword.

ALL_KEYWORDS

Array of all SQL keywords as string constants.

ALL_KEYWORDS_INDEX

Array of all Keyword enum values in declaration order.

ALTER

The ALTER SQL keyword.

ALWAYS

The ALWAYS SQL keyword.

ANALYZE

The ANALYZE SQL keyword.

AND

The AND SQL keyword.

ANTI

The ANTI SQL keyword.

ANY

The ANY SQL keyword.

APPLICATION

The APPLICATION SQL keyword.

APPLY

The APPLY SQL keyword.

APPLYBUDGET

The APPLYBUDGET SQL keyword.

ARCHIVE

The ARCHIVE SQL keyword.

ARE

The ARE SQL keyword.

ARRAY

The ARRAY SQL keyword.

ARRAY_MAX_CARDINALITY

The ARRAY_MAX_CARDINALITY SQL keyword.

AS

The AS SQL keyword.

ASC

The ASC SQL keyword.

ASENSITIVE

The ASENSITIVE SQL keyword.

ASOF

The ASOF SQL keyword.

ASSERT

The ASSERT SQL keyword.

ASYMMETRIC

The ASYMMETRIC SQL keyword.

AT

The AT SQL keyword.

ATOMIC

The ATOMIC SQL keyword.

ATTACH

The ATTACH SQL keyword.

AUDIT

The AUDIT SQL keyword.

AUTHENTICATION

The AUTHENTICATION SQL keyword.

AUTHORIZATION

The AUTHORIZATION SQL keyword.

AUTHORIZATIONS

The AUTHORIZATIONS SQL keyword.

AUTO

The AUTO SQL keyword.

AUTOEXTEND_SIZE

The AUTOEXTEND_SIZE SQL keyword.

AUTOINCREMENT

The AUTOINCREMENT SQL keyword.

AUTO_INCREMENT

The AUTO_INCREMENT SQL keyword.

AVG

The AVG SQL keyword.

AVG_ROW_LENGTH

The AVG_ROW_LENGTH SQL keyword.

AVRO

The AVRO SQL keyword.

BACKWARD

The BACKWARD SQL keyword.

BASE64

The BASE64 SQL keyword.

BASE_LOCATION

The BASE_LOCATION SQL keyword.

BEFORE

The BEFORE SQL keyword.

BEGIN

The BEGIN SQL keyword.

BEGIN_FRAME

The BEGIN_FRAME SQL keyword.

BEGIN_PARTITION

The BEGIN_PARTITION SQL keyword.

BERNOULLI

The BERNOULLI SQL keyword.

BETWEEN

The BETWEEN SQL keyword.

BIGDECIMAL

The BIGDECIMAL SQL keyword.

BIGINT

The BIGINT SQL keyword.

BIGNUMERIC

The BIGNUMERIC SQL keyword.

BINARY

The BINARY SQL keyword.

BIND

The BIND SQL keyword.

BINDING

The BINDING SQL keyword.

BIT

The BIT SQL keyword.

BLANKSASNULL

The BLANKSASNULL SQL keyword.

BLOB

The BLOB SQL keyword.

BLOCK

The BLOCK SQL keyword.

BLOOM

The BLOOM SQL keyword.

BLOOMFILTER

The BLOOMFILTER SQL keyword.

BOOL

The BOOL SQL keyword.

BOOLEAN

The BOOLEAN SQL keyword.

BOOST

The BOOST SQL keyword.

BOTH

The BOTH SQL keyword.

BOX

The BOX SQL keyword.

BRIN

The BRIN SQL keyword.

BROWSE

The BROWSE SQL keyword.

BTREE

The BTREE SQL keyword.

BUCKET

The BUCKET SQL keyword.

BUCKETS

The BUCKETS SQL keyword.

BY

The BY SQL keyword.

BYPASSRLS

The BYPASSRLS SQL keyword.

BYTEA

The BYTEA SQL keyword.

BYTES

The BYTES SQL keyword.

BZIP2

The BZIP2 SQL keyword.

CACHE

The CACHE SQL keyword.

CALL

The CALL SQL keyword.

CALLED

The CALLED SQL keyword.

CANONICAL

The CANONICAL SQL keyword.

CARDINALITY

The CARDINALITY SQL keyword.

CASCADE

The CASCADE SQL keyword.

CASCADED

The CASCADED SQL keyword.

CASE

The CASE SQL keyword.

CASES

The CASES SQL keyword.

CAST

The CAST SQL keyword.

CATALOG

The CATALOG SQL keyword.

CATALOG_SYNC

The CATALOG_SYNC SQL keyword.

CATALOG_SYNC_NAMESPACE_FLATTEN_DELIMITER

The CATALOG_SYNC_NAMESPACE_FLATTEN_DELIMITER SQL keyword.

CATALOG_SYNC_NAMESPACE_MODE

The CATALOG_SYNC_NAMESPACE_MODE SQL keyword.

CATCH

The CATCH SQL keyword.

CATEGORY

The CATEGORY SQL keyword.

CEIL

The CEIL SQL keyword.

CEILING

The CEILING SQL keyword.

CENTURY

The CENTURY SQL keyword.

CHAIN

The CHAIN SQL keyword.

CHANGE

The CHANGE SQL keyword.

CHANGE_TRACKING

The CHANGE_TRACKING SQL keyword.

CHANNEL

The CHANNEL SQL keyword.

CHAR

The CHAR SQL keyword.

CHARACTER

The CHARACTER SQL keyword.

CHARACTERISTICS

The CHARACTERISTICS SQL keyword.

CHARACTERS

The CHARACTERS SQL keyword.

CHARACTER_LENGTH

The CHARACTER_LENGTH SQL keyword.

CHARSET

The CHARSET SQL keyword.

CHAR_LENGTH

The CHAR_LENGTH SQL keyword.

CHECK

The CHECK SQL keyword.

CHECKSUM

The CHECKSUM SQL keyword.

CIRCLE

The CIRCLE SQL keyword.

CLASS

The CLASS SQL keyword.

CLEANPATH

The CLEANPATH SQL keyword.

CLEAR

The CLEAR SQL keyword.

CLOB

The CLOB SQL keyword.

CLONE

The CLONE SQL keyword.

CLOSE

The CLOSE SQL keyword.

CLUSTER

The CLUSTER SQL keyword.

CLUSTERED

The CLUSTERED SQL keyword.

CLUSTERING

The CLUSTERING SQL keyword.

COALESCE

The COALESCE SQL keyword.

COLLATABLE

The COLLATABLE SQL keyword.

COLLATE

The COLLATE SQL keyword.

COLLATION

The COLLATION SQL keyword.

COLLECT

The COLLECT SQL keyword.

COLLECTION

The COLLECTION SQL keyword.

COLUMN

The COLUMN SQL keyword.

COLUMNS

The COLUMNS SQL keyword.

COLUMNSTORE

The COLUMNSTORE SQL keyword.

COMMENT

The COMMENT SQL keyword.

COMMIT

The COMMIT SQL keyword.

COMMITTED

The COMMITTED SQL keyword.

COMMUTATOR

The COMMUTATOR SQL keyword.

COMPATIBLE

The COMPATIBLE SQL keyword.

COMPRESS

The COMPRESS SQL keyword.

COMPRESSION

The COMPRESSION SQL keyword.

COMPUPDATE

The COMPUPDATE SQL keyword.

COMPUTE

The COMPUTE SQL keyword.

CONCURRENTLY

The CONCURRENTLY SQL keyword.

CONDITION

The CONDITION SQL keyword.

CONFLICT

The CONFLICT SQL keyword.

CONNECT

The CONNECT SQL keyword.

CONNECTION

The CONNECTION SQL keyword.

CONNECTOR

The CONNECTOR SQL keyword.

CONNECT_BY_ROOT

The CONNECT_BY_ROOT SQL keyword.

CONSTRAINT

The CONSTRAINT SQL keyword.

CONTACT

The CONTACT SQL keyword.

CONTAINS

The CONTAINS SQL keyword.

CONTINUE

The CONTINUE SQL keyword.

CONVERT

The CONVERT SQL keyword.

COPY

The COPY SQL keyword.

COPY_OPTIONS

The COPY_OPTIONS SQL keyword.

CORR

The CORR SQL keyword.

CORRESPONDING

The CORRESPONDING SQL keyword.

COUNT

The COUNT SQL keyword.

COVAR_POP

The COVAR_POP SQL keyword.

COVAR_SAMP

The COVAR_SAMP SQL keyword.

CREATE

The CREATE SQL keyword.

CREATEDB

The CREATEDB SQL keyword.

CREATEROLE

The CREATEROLE SQL keyword.

CREDENTIALS

The CREDENTIALS SQL keyword.

CROSS

The CROSS SQL keyword.

CSV

The CSV SQL keyword.

CUBE

The CUBE SQL keyword.

CUME_DIST

The CUME_DIST SQL keyword.

CURRENT

The CURRENT SQL keyword.

CURRENT_CATALOG

The CURRENT_CATALOG SQL keyword.

CURRENT_DATE

The CURRENT_DATE SQL keyword.

CURRENT_DEFAULT_TRANSFORM_GROUP

The CURRENT_DEFAULT_TRANSFORM_GROUP SQL keyword.

CURRENT_PATH

The CURRENT_PATH SQL keyword.

CURRENT_ROLE

The CURRENT_ROLE SQL keyword.

CURRENT_ROW

The CURRENT_ROW SQL keyword.

CURRENT_SCHEMA

The CURRENT_SCHEMA SQL keyword.

CURRENT_TIME

The CURRENT_TIME SQL keyword.

CURRENT_TIMESTAMP

The CURRENT_TIMESTAMP SQL keyword.

CURRENT_TRANSFORM_GROUP_FOR_TYPE

The CURRENT_TRANSFORM_GROUP_FOR_TYPE SQL keyword.

CURRENT_USER

The CURRENT_USER SQL keyword.

CURSOR

The CURSOR SQL keyword.

CYCLE

The CYCLE SQL keyword.

DATA

The DATA SQL keyword.

DATABASE

The DATABASE SQL keyword.

DATABASES

The DATABASES SQL keyword.

DATA_RETENTION_TIME_IN_DAYS

The DATA_RETENTION_TIME_IN_DAYS SQL keyword.

DATE

The DATE SQL keyword.

DATE32

The DATE32 SQL keyword.

DATEFORMAT

The DATEFORMAT SQL keyword.

DATETIME

The DATETIME SQL keyword.

DATETIME64

The DATETIME64 SQL keyword.

DAY

The DAY SQL keyword.

DAYOFWEEK

The DAYOFWEEK SQL keyword.

DAYOFYEAR

The DAYOFYEAR SQL keyword.

DAYS

The DAYS SQL keyword.

DCPROPERTIES

The DCPROPERTIES SQL keyword.

DEALLOCATE

The DEALLOCATE SQL keyword.

DEC

The DEC SQL keyword.

DECADE

The DECADE SQL keyword.

DECIMAL

The DECIMAL SQL keyword.

DECLARE

The DECLARE SQL keyword.

DEDUPLICATE

The DEDUPLICATE SQL keyword.

DEFAULT

The DEFAULT SQL keyword.

DEFAULTS

The DEFAULTS SQL keyword.

DEFAULT_DDL_COLLATION

The DEFAULT_DDL_COLLATION SQL keyword.

DEFAULT_MFA_METHOD

The DEFAULT_MFA_METHOD SQL keyword.

DEFAULT_SECONDARY_ROLES

The DEFAULT_SECONDARY_ROLES SQL keyword.

DEFERRABLE

The DEFERRABLE SQL keyword.

DEFERRED

The DEFERRED SQL keyword.

DEFINE

The DEFINE SQL keyword.

DEFINED

The DEFINED SQL keyword.

DEFINER

The DEFINER SQL keyword.

DELAYED

The DELAYED SQL keyword.

DELAY_KEY_WRITE

The DELAY_KEY_WRITE SQL keyword.

DELEGATED

The DELEGATED SQL keyword.

DELETE

The DELETE SQL keyword.

DELIMITED

The DELIMITED SQL keyword.

DELIMITER

The DELIMITER SQL keyword.

DELTA

The DELTA SQL keyword.

DENSE_RANK

The DENSE_RANK SQL keyword.

DENY

The DENY SQL keyword.

DEREF

The DEREF SQL keyword.

DESC

The DESC SQL keyword.

DESCRIBE

The DESCRIBE SQL keyword.

DETACH

The DETACH SQL keyword.

DETAIL

The DETAIL SQL keyword.

DETERMINISTIC

The DETERMINISTIC SQL keyword.

DIMENSIONS

The DIMENSIONS SQL keyword.

DIRECTORY

The DIRECTORY SQL keyword.

DISABLE

The DISABLE SQL keyword.

DISCARD

The DISCARD SQL keyword.

DISCONNECT

The DISCONNECT SQL keyword.

DISTINCT

The DISTINCT SQL keyword.

DISTINCTROW

The DISTINCTROW SQL keyword.

DISTRIBUTE

The DISTRIBUTE SQL keyword.

DIV

The DIV SQL keyword.

DO

The DO SQL keyword.

DOMAIN

The DOMAIN SQL keyword.

DOUBLE

The DOUBLE SQL keyword.

DOW

The DOW SQL keyword.

DOWNSTREAM

The DOWNSTREAM SQL keyword.

DOY

The DOY SQL keyword.

DROP

The DROP SQL keyword.

DRY

The DRY SQL keyword.

DUO

The DUO SQL keyword.

DUPLICATE

The DUPLICATE SQL keyword.

DYNAMIC

The DYNAMIC SQL keyword.

EACH

The EACH SQL keyword.

ELEMENT

The ELEMENT SQL keyword.

ELEMENTS

The ELEMENTS SQL keyword.

ELSE

The ELSE SQL keyword.

ELSEIF

The ELSEIF SQL keyword.

EMPTY

The EMPTY SQL keyword.

EMPTYASNULL

The EMPTYASNULL SQL keyword.

ENABLE

The ENABLE SQL keyword.

ENABLE_SCHEMA_EVOLUTION

The ENABLE_SCHEMA_EVOLUTION SQL keyword.

ENCODING

The ENCODING SQL keyword.

ENCRYPTED

The ENCRYPTED SQL keyword.

ENCRYPTION

The ENCRYPTION SQL keyword.

END

The END SQL keyword.

ENDPOINT

The ENDPOINT SQL keyword.

END_EXEC

The END-EXEC SQL keyword.

END_FRAME

The END_FRAME SQL keyword.

END_PARTITION

The END_PARTITION SQL keyword.

ENFORCED

The ENFORCED SQL keyword.

ENGINE

The ENGINE SQL keyword.

ENGINE_ATTRIBUTE

The ENGINE_ATTRIBUTE SQL keyword.

ENROLL

The ENROLL SQL keyword.

ENUM

The ENUM SQL keyword.

ENUM8

The ENUM8 SQL keyword.

ENUM16

The ENUM16 SQL keyword.

EPHEMERAL

The EPHEMERAL SQL keyword.

EPOCH

The EPOCH SQL keyword.

EQUALS

The EQUALS SQL keyword.

ERROR

The ERROR SQL keyword.

ESCAPE

The ESCAPE SQL keyword.

ESCAPED

The ESCAPED SQL keyword.

ESTIMATE

The ESTIMATE SQL keyword.

EVENT

The EVENT SQL keyword.

EVERY

The EVERY SQL keyword.

EVOLVE

The EVOLVE SQL keyword.

EXCEPT

The EXCEPT SQL keyword.

EXCEPTION

The EXCEPTION SQL keyword.

EXCHANGE

The EXCHANGE SQL keyword.

EXCLUDE

The EXCLUDE SQL keyword.

EXCLUDING

The EXCLUDING SQL keyword.

EXCLUSIVE

The EXCLUSIVE SQL keyword.

EXEC

The EXEC SQL keyword.

EXECUTE

The EXECUTE SQL keyword.

EXECUTION

The EXECUTION SQL keyword.

EXISTS

The EXISTS SQL keyword.

EXP

The EXP SQL keyword.

EXPANSION

The EXPANSION SQL keyword.

EXPLAIN

The EXPLAIN SQL keyword.

EXPLICIT

The EXPLICIT SQL keyword.

EXPORT

The EXPORT SQL keyword.

EXTEND

The EXTEND SQL keyword.

EXTENDED

The EXTENDED SQL keyword.

EXTENSION

The EXTENSION SQL keyword.

EXTERNAL

The EXTERNAL SQL keyword.

EXTERNAL_VOLUME

The EXTERNAL_VOLUME SQL keyword.

EXTRACT

The EXTRACT SQL keyword.

FACTS

The FACTS SQL keyword.

FAIL

The FAIL SQL keyword.

FAILOVER

The FAILOVER SQL keyword.

FALSE

The FALSE SQL keyword.

FAMILY

The FAMILY SQL keyword.

FETCH

The FETCH SQL keyword.

FIELDS

The FIELDS SQL keyword.

FILE

The FILE SQL keyword.

FILES

The FILES SQL keyword.

FILE_FORMAT

The FILE_FORMAT SQL keyword.

FILL

The FILL SQL keyword.

FILTER

The FILTER SQL keyword.

FINAL

The FINAL SQL keyword.

FIRST

The FIRST SQL keyword.

FIRST_VALUE

The FIRST_VALUE SQL keyword.

FIXEDSTRING

The FIXEDSTRING SQL keyword.

FIXEDWIDTH

The FIXEDWIDTH SQL keyword.

FLATTEN

The FLATTEN SQL keyword.

FLOAT

The FLOAT SQL keyword.

FLOAT4

The FLOAT4 SQL keyword.

FLOAT8

The FLOAT8 SQL keyword.

FLOAT32

The FLOAT32 SQL keyword.

FLOAT64

The FLOAT64 SQL keyword.

FLOOR

The FLOOR SQL keyword.

FLUSH

The FLUSH SQL keyword.

FN

The FN SQL keyword.

FOLLOWING

The FOLLOWING SQL keyword.

FOR

The FOR SQL keyword.

FORCE

The FORCE SQL keyword.

FORCE_NOT_NULL

The FORCE_NOT_NULL SQL keyword.

FORCE_NULL

The FORCE_NULL SQL keyword.

FORCE_QUOTE

The FORCE_QUOTE SQL keyword.

FOREIGN

The FOREIGN SQL keyword.

FORMAT

The FORMAT SQL keyword.

FORMATTED

The FORMATTED SQL keyword.

FORWARD

The FORWARD SQL keyword.

FRAME_ROW

The FRAME_ROW SQL keyword.

FREE

The FREE SQL keyword.

FREEZE

The FREEZE SQL keyword.

FROM

The FROM SQL keyword.

FSCK

The FSCK SQL keyword.

FULFILLMENT

The FULFILLMENT SQL keyword.

FULL

The FULL SQL keyword.

FULLTEXT

The FULLTEXT SQL keyword.

FUNCTION

The FUNCTION SQL keyword.

FUNCTIONS

The FUNCTIONS SQL keyword.

FUSION

The FUSION SQL keyword.

FUTURE

The FUTURE SQL keyword.

GB

The GB SQL keyword.

GENERAL

The GENERAL SQL keyword.

GENERATE

The GENERATE SQL keyword.

GENERATED

The GENERATED SQL keyword.

GEOGRAPHY

The GEOGRAPHY SQL keyword.

GET

The GET SQL keyword.

GIN

The GIN SQL keyword.

GIST

The GIST SQL keyword.

GLOBAL

The GLOBAL SQL keyword.

GRANT

The GRANT SQL keyword.

GRANTED

The GRANTED SQL keyword.

GRANTS

The GRANTS SQL keyword.

GRAPHVIZ

The GRAPHVIZ SQL keyword.

GROUP

The GROUP SQL keyword.

GROUPING

The GROUPING SQL keyword.

GROUPS

The GROUPS SQL keyword.

GZIP

The GZIP SQL keyword.

HASH

The HASH SQL keyword.

HASHES

The HASHES SQL keyword.

HAVING

The HAVING SQL keyword.

HEADER

The HEADER SQL keyword.

HEAP

The HEAP SQL keyword.

HIGH_PRIORITY

The HIGH_PRIORITY SQL keyword.

HISTORY

The HISTORY SQL keyword.

HIVEVAR

The HIVEVAR SQL keyword.

HOLD

The HOLD SQL keyword.

HOSTS

The HOSTS SQL keyword.

HOUR

The HOUR SQL keyword.

HOURS

The HOURS SQL keyword.

HUGEINT

The HUGEINT SQL keyword.

IAM_ROLE

The IAM_ROLE SQL keyword.

ICEBERG

The ICEBERG SQL keyword.

ID

The ID SQL keyword.

IDENTIFIED

The IDENTIFIED SQL keyword.

IDENTITY

The IDENTITY SQL keyword.

IDENTITY_INSERT

The IDENTITY_INSERT SQL keyword.

IF

The IF SQL keyword.

IGNORE

The IGNORE SQL keyword.

IGNOREHEADER

The IGNOREHEADER SQL keyword.

ILIKE

The ILIKE SQL keyword.

IMMEDIATE

The IMMEDIATE SQL keyword.

IMMUTABLE

The IMMUTABLE SQL keyword.

IMPORT

The IMPORT SQL keyword.

IMPORTED

The IMPORTED SQL keyword.

IN

The IN SQL keyword.

INCLUDE

The INCLUDE SQL keyword.

INCLUDE_NULL_VALUES

The INCLUDE_NULL_VALUES SQL keyword.

INCLUDING

The INCLUDING SQL keyword.

INCREMENT

The INCREMENT SQL keyword.

INCREMENTAL

The INCREMENTAL SQL keyword.

INDEX

The INDEX SQL keyword.

INDICATOR

The INDICATOR SQL keyword.

INHERIT

The INHERIT SQL keyword.

INHERITS

The INHERITS SQL keyword.

INITIALIZE

The INITIALIZE SQL keyword.

INITIALLY

The INITIALLY SQL keyword.

INNER

The INNER SQL keyword.

INOUT

The INOUT SQL keyword.

INPATH

The INPATH SQL keyword.

INPLACE

The INPLACE SQL keyword.

INPUT

The INPUT SQL keyword.

INPUTFORMAT

The INPUTFORMAT SQL keyword.

INSENSITIVE

The INSENSITIVE SQL keyword.

INSERT

The INSERT SQL keyword.

INSERT_METHOD

The INSERT_METHOD SQL keyword.

INSTALL

The INSTALL SQL keyword.

INSTANT

The INSTANT SQL keyword.

INSTEAD

The INSTEAD SQL keyword.

INT

The INT SQL keyword.

INT2

The INT2 SQL keyword.

INT4

The INT4 SQL keyword.

INT8

The INT8 SQL keyword.

INT16

The INT16 SQL keyword.

INT32

The INT32 SQL keyword.

INT64

The INT64 SQL keyword.

INT128

The INT128 SQL keyword.

INT256

The INT256 SQL keyword.

INTEGER

The INTEGER SQL keyword.

INTEGRATION

The INTEGRATION SQL keyword.

INTERNALLENGTH

The INTERNALLENGTH SQL keyword.

INTERPOLATE

The INTERPOLATE SQL keyword.

INTERSECT

The INTERSECT SQL keyword.

INTERSECTION

The INTERSECTION SQL keyword.

INTERVAL

The INTERVAL SQL keyword.

INTO

The INTO SQL keyword.

INVISIBLE

The INVISIBLE SQL keyword.

INVOKER

The INVOKER SQL keyword.

IO

The IO SQL keyword.

IS

The IS SQL keyword.

ISODOW

The ISODOW SQL keyword.

ISOLATION

The ISOLATION SQL keyword.

ISOWEEK

The ISOWEEK SQL keyword.

ISOYEAR

The ISOYEAR SQL keyword.

ITEMS

The ITEMS SQL keyword.

JAR

The JAR SQL keyword.

JOIN

The JOIN SQL keyword.

JSON

The JSON SQL keyword.

JSONB

The JSONB SQL keyword.

JSONFILE

The JSONFILE SQL keyword.

JSON_TABLE

The JSON_TABLE SQL keyword.

JULIAN

The JULIAN SQL keyword.

KEY

The KEY SQL keyword.

KEYS

The KEYS SQL keyword.

KEY_BLOCK_SIZE

The KEY_BLOCK_SIZE SQL keyword.

KILL

The KILL SQL keyword.

LAG

The LAG SQL keyword.

LAMBDA

The LAMBDA SQL keyword.

LANGUAGE

The LANGUAGE SQL keyword.

LARGE

The LARGE SQL keyword.

LAST

The LAST SQL keyword.

LAST_VALUE

The LAST_VALUE SQL keyword.

LATERAL

The LATERAL SQL keyword.

LEAD

The LEAD SQL keyword.

LEADING

The LEADING SQL keyword.

LEAKPROOF

The LEAKPROOF SQL keyword.

LEAST

The LEAST SQL keyword.

LEFT

The LEFT SQL keyword.

LEFTARG

The LEFTARG SQL keyword.

LEVEL

The LEVEL SQL keyword.

LIKE

The LIKE SQL keyword.

LIKE_REGEX

The LIKE_REGEX SQL keyword.

LIMIT

The LIMIT SQL keyword.

LINE

The LINE SQL keyword.

LINES

The LINES SQL keyword.

LIST

The LIST SQL keyword.

LISTEN

The LISTEN SQL keyword.

LISTING

The LISTING SQL keyword.

LN

The LN SQL keyword.

LOAD

The LOAD SQL keyword.

LOCAL

The LOCAL SQL keyword.

LOCALTIME

The LOCALTIME SQL keyword.

LOCALTIMESTAMP

The LOCALTIMESTAMP SQL keyword.

LOCATION

The LOCATION SQL keyword.

LOCK

The LOCK SQL keyword.

LOCKED

The LOCKED SQL keyword.

LOG

The LOG SQL keyword.

LOGIN

The LOGIN SQL keyword.

LOGS

The LOGS SQL keyword.

LONG

The LONG SQL keyword.

LONGBLOB

The LONGBLOB SQL keyword.

LONGTEXT

The LONGTEXT SQL keyword.

LOWCARDINALITY

The LOWCARDINALITY SQL keyword.

LOWER

The LOWER SQL keyword.

LOW_PRIORITY

The LOW_PRIORITY SQL keyword.

LS

The LS SQL keyword.

LSEG

The LSEG SQL keyword.

MACRO

The MACRO SQL keyword.

MAIN

The MAIN SQL keyword.

MANAGE

The MANAGE SQL keyword.

MANAGED

The MANAGED SQL keyword.

MANAGEDLOCATION

The MANAGEDLOCATION SQL keyword.

MANIFEST

The MANIFEST SQL keyword.

MAP

The MAP SQL keyword.

MASKING

The MASKING SQL keyword.

MATCH

The MATCH SQL keyword.

MATCHED

The MATCHED SQL keyword.

MATCHES

The MATCHES SQL keyword.

MATCH_CONDITION

The MATCH_CONDITION SQL keyword.

MATCH_RECOGNIZE

The MATCH_RECOGNIZE SQL keyword.

MATERIALIZE

The MATERIALIZE SQL keyword.

MATERIALIZED

The MATERIALIZED SQL keyword.

MAX

The MAX SQL keyword.

MAXFILESIZE

The MAXFILESIZE SQL keyword.

MAXVALUE

The MAXVALUE SQL keyword.

MAX_DATA_EXTENSION_TIME_IN_DAYS

The MAX_DATA_EXTENSION_TIME_IN_DAYS SQL keyword.

MAX_ROWS

The MAX_ROWS SQL keyword.

MB

The MB SQL keyword.

MEASURES

The MEASURES SQL keyword.

MEDIUMBLOB

The MEDIUMBLOB SQL keyword.

MEDIUMINT

The MEDIUMINT SQL keyword.

MEDIUMTEXT

The MEDIUMTEXT SQL keyword.

MEMBER

The MEMBER SQL keyword.

MERGE

The MERGE SQL keyword.

MERGES

The MERGES SQL keyword.

MESSAGE

The MESSAGE SQL keyword.

METADATA

The METADATA SQL keyword.

METHOD

The METHOD SQL keyword.

METRIC

The METRIC SQL keyword.

METRICS

The METRICS SQL keyword.

MFA

The MFA SQL keyword.

MICROSECOND

The MICROSECOND SQL keyword.

MICROSECONDS

The MICROSECONDS SQL keyword.

MILLENIUM

The MILLENIUM SQL keyword.

MILLENNIUM

The MILLENNIUM SQL keyword.

MILLISECOND

The MILLISECOND SQL keyword.

MILLISECONDS

The MILLISECONDS SQL keyword.

MIN

The MIN SQL keyword.

MINUS

The MINUS SQL keyword.

MINUTE

The MINUTE SQL keyword.

MINUTES

The MINUTES SQL keyword.

MINVALUE

The MINVALUE SQL keyword.

MIN_ROWS

The MIN_ROWS SQL keyword.

MOD

The MOD SQL keyword.

MODE

The MODE SQL keyword.

MODIFIES

The MODIFIES SQL keyword.

MODIFY

The MODIFY SQL keyword.

MODULE

The MODULE SQL keyword.

MODULUS

The MODULUS SQL keyword.

MONITOR

The MONITOR SQL keyword.

MONTH

The MONTH SQL keyword.

MONTHS

The MONTHS SQL keyword.

MSCK

The MSCK SQL keyword.

MULTIRANGE_TYPE_NAME

The MULTIRANGE_TYPE_NAME SQL keyword.

MULTISET

The MULTISET SQL keyword.

MUTATION

The MUTATION SQL keyword.

NAME

The NAME SQL keyword.

NAMES

The NAMES SQL keyword.

NANOSECOND

The NANOSECOND SQL keyword.

NANOSECONDS

The NANOSECONDS SQL keyword.

NATIONAL

The NATIONAL SQL keyword.

NATURAL

The NATURAL SQL keyword.

NCHAR

The NCHAR SQL keyword.

NCLOB

The NCLOB SQL keyword.

NEGATOR

The NEGATOR SQL keyword.

NEST

The NEST SQL keyword.

NESTED

The NESTED SQL keyword.

NETWORK

The NETWORK SQL keyword.

NEW

The NEW SQL keyword.

NEXT

The NEXT SQL keyword.

NFC

The NFC SQL keyword.

NFD

The NFD SQL keyword.

NFKC

The NFKC SQL keyword.

NFKD

The NFKD SQL keyword.

NO

The NO SQL keyword.

NOBYPASSRLS

The NOBYPASSRLS SQL keyword.

NOCOMPRESS

The NOCOMPRESS SQL keyword.

NOCREATEDB

The NOCREATEDB SQL keyword.

NOCREATEROLE

The NOCREATEROLE SQL keyword.

NOCYCLE

The NOCYCLE SQL keyword.

NOINHERIT

The NOINHERIT SQL keyword.

NOLOGIN

The NOLOGIN SQL keyword.

NONE

The NONE SQL keyword.

NOORDER

The NOORDER SQL keyword.

NOREPLICATION

The NOREPLICATION SQL keyword.

NORMALIZE

The NORMALIZE SQL keyword.

NORMALIZED

The NORMALIZED SQL keyword.

NOSCAN

The NOSCAN SQL keyword.

NOSUPERUSER

The NOSUPERUSER SQL keyword.

NOT

The NOT SQL keyword.

NOTHING

The NOTHING SQL keyword.

NOTIFY

The NOTIFY SQL keyword.

NOTNULL

The NOTNULL SQL keyword.

NOWAIT

The NOWAIT SQL keyword.

NO_WRITE_TO_BINLOG

The NO_WRITE_TO_BINLOG SQL keyword.

NTH_VALUE

The NTH_VALUE SQL keyword.

NTILE

The NTILE SQL keyword.

NULL

The NULL SQL keyword.

NULLABLE

The NULLABLE SQL keyword.

NULLIF

The NULLIF SQL keyword.

NULLS

The NULLS SQL keyword.

NUMBER

The NUMBER SQL keyword.

NUMERIC

The NUMERIC SQL keyword.

NVARCHAR

The NVARCHAR SQL keyword.

OBJECT

The OBJECT SQL keyword.

OBJECTS

The OBJECTS SQL keyword.

OCCURRENCES_REGEX

The OCCURRENCES_REGEX SQL keyword.

OCTETS

The OCTETS SQL keyword.

OCTET_LENGTH

The OCTET_LENGTH SQL keyword.

OF

The OF SQL keyword.

OFF

The OFF SQL keyword.

OFFSET

The OFFSET SQL keyword.

OFFSETS

The OFFSETS SQL keyword.

OLD

The OLD SQL keyword.

OMIT

The OMIT SQL keyword.

ON

The ON SQL keyword.

ONE

The ONE SQL keyword.

ONLY

The ONLY SQL keyword.

ON_CREATE

The ON_CREATE SQL keyword.

ON_SCHEDULE

The ON_SCHEDULE SQL keyword.

OPEN

The OPEN SQL keyword.

OPENJSON

The OPENJSON SQL keyword.

OPERATE

The OPERATE SQL keyword.

OPERATOR

The OPERATOR SQL keyword.

OPTIMIZATION

The OPTIMIZATION SQL keyword.

OPTIMIZE

The OPTIMIZE SQL keyword.

OPTIMIZED

The OPTIMIZED SQL keyword.

OPTIMIZER_COSTS

The OPTIMIZER_COSTS SQL keyword.

OPTION

The OPTION SQL keyword.

OPTIONS

The OPTIONS SQL keyword.

OR

The OR SQL keyword.

ORC

The ORC SQL keyword.

ORDER

The ORDER SQL keyword.

ORDINALITY

The ORDINALITY SQL keyword.

ORGANIZATION

The ORGANIZATION SQL keyword.

OTHER

The OTHER SQL keyword.

OTP

The OTP SQL keyword.

OUT

The OUT SQL keyword.

OUTER

The OUTER SQL keyword.

OUTPUT

The OUTPUT SQL keyword.

OUTPUTFORMAT

The OUTPUTFORMAT SQL keyword.

OVER

The OVER SQL keyword.

OVERFLOW

The OVERFLOW SQL keyword.

OVERLAPS

The OVERLAPS SQL keyword.

OVERLAY

The OVERLAY SQL keyword.

OVERRIDE

The OVERRIDE SQL keyword.

OVERWRITE

The OVERWRITE SQL keyword.

OWNED

The OWNED SQL keyword.

OWNER

The OWNER SQL keyword.

OWNERSHIP

The OWNERSHIP SQL keyword.

PACKAGE

The PACKAGE SQL keyword.

PACKAGES

The PACKAGES SQL keyword.

PACK_KEYS

The PACK_KEYS SQL keyword.

PARALLEL

The PARALLEL SQL keyword.

PARAMETER

The PARAMETER SQL keyword.

PARQUET

The PARQUET SQL keyword.

PART

The PART SQL keyword.

PARTIAL

The PARTIAL SQL keyword.

PARTITION

The PARTITION SQL keyword.

PARTITIONED

The PARTITIONED SQL keyword.

PARTITIONS

The PARTITIONS SQL keyword.

PASSEDBYVALUE

The PASSEDBYVALUE SQL keyword.

PASSING

The PASSING SQL keyword.

PASSKEY

The PASSKEY SQL keyword.

PASSWORD

The PASSWORD SQL keyword.

PAST

The PAST SQL keyword.

PATH

The PATH SQL keyword.

PATTERN

The PATTERN SQL keyword.

PCTFREE

The PCTFREE SQL keyword.

PER

The PER SQL keyword.

PERCENT

The PERCENT SQL keyword.

PERCENTILE_CONT

The PERCENTILE_CONT SQL keyword.

PERCENTILE_DISC

The PERCENTILE_DISC SQL keyword.

PERCENT_RANK

The PERCENT_RANK SQL keyword.

PERIOD

The PERIOD SQL keyword.

PERMISSIVE

The PERMISSIVE SQL keyword.

PERSISTENT

The PERSISTENT SQL keyword.

PIVOT

The PIVOT SQL keyword.

PLACING

The PLACING SQL keyword.

PLAIN

The PLAIN SQL keyword.

PLAN

The PLAN SQL keyword.

PLANS

The PLANS SQL keyword.

POINT

The POINT SQL keyword.

POLICY

The POLICY SQL keyword.

POLYGON

The POLYGON SQL keyword.

POOL

The POOL SQL keyword.

PORTION

The PORTION SQL keyword.

POSITION

The POSITION SQL keyword.

POSITION_REGEX

The POSITION_REGEX SQL keyword.

POWER

The POWER SQL keyword.

PRAGMA

The PRAGMA SQL keyword.

PRECEDES

The PRECEDES SQL keyword.

PRECEDING

The PRECEDING SQL keyword.

PRECISION

The PRECISION SQL keyword.

PREFERRED

The PREFERRED SQL keyword.

PREPARE

The PREPARE SQL keyword.

PRESERVE

The PRESERVE SQL keyword.

PRESET

The PRESET SQL keyword.

PREWHERE

The PREWHERE SQL keyword.

PRIMARY

The PRIMARY SQL keyword.

PRINT

The PRINT SQL keyword.

PRIOR

The PRIOR SQL keyword.

PRIVILEGES

The PRIVILEGES SQL keyword.

PROCEDURE

The PROCEDURE SQL keyword.

PROFILE

The PROFILE SQL keyword.

PROGRAM

The PROGRAM SQL keyword.

PROJECTION

The PROJECTION SQL keyword.

PUBLIC

The PUBLIC SQL keyword.

PURCHASE

The PURCHASE SQL keyword.

PURGE

The PURGE SQL keyword.

QUALIFY

The QUALIFY SQL keyword.

QUARTER

The QUARTER SQL keyword.

QUERIES

The QUERIES SQL keyword.

QUERY

The QUERY SQL keyword.

QUOTE

The QUOTE SQL keyword.

RAISE

The RAISE SQL keyword.

RAISERROR

The RAISERROR SQL keyword.

RANGE

The RANGE SQL keyword.

RANK

The RANK SQL keyword.

RAW

The RAW SQL keyword.

RCFILE

The RCFILE SQL keyword.

READ

The READ SQL keyword.

READS

The READS SQL keyword.

READ_ONLY

The READ_ONLY SQL keyword.

REAL

The REAL SQL keyword.

RECEIVE

The RECEIVE SQL keyword.

RECLUSTER

The RECLUSTER SQL keyword.

RECURSIVE

The RECURSIVE SQL keyword.

REF

The REF SQL keyword.

REFERENCES

The REFERENCES SQL keyword.

REFERENCING

The REFERENCING SQL keyword.

REFRESH

The REFRESH SQL keyword.

REFRESH_MODE

The REFRESH_MODE SQL keyword.

REGCLASS

The REGCLASS SQL keyword.

REGEXP

The REGEXP SQL keyword.

REGION

The REGION SQL keyword.

REGR_AVGX

The REGR_AVGX SQL keyword.

REGR_AVGY

The REGR_AVGY SQL keyword.

REGR_COUNT

The REGR_COUNT SQL keyword.

REGR_INTERCEPT

The REGR_INTERCEPT SQL keyword.

REGR_R2

The REGR_R2 SQL keyword.

REGR_SLOPE

The REGR_SLOPE SQL keyword.

REGR_SXX

The REGR_SXX SQL keyword.

REGR_SXY

The REGR_SXY SQL keyword.

REGR_SYY

The REGR_SYY SQL keyword.

REINDEX

The REINDEX SQL keyword.

RELATIVE

The RELATIVE SQL keyword.

RELAY

The RELAY SQL keyword.

RELEASE

The RELEASE SQL keyword.

RELEASES

The RELEASES SQL keyword.

REMAINDER

The REMAINDER SQL keyword.

REMOTE

The REMOTE SQL keyword.

REMOVE

The REMOVE SQL keyword.

REMOVEQUOTES

The REMOVEQUOTES SQL keyword.

RENAME

The RENAME SQL keyword.

REORG

The REORG SQL keyword.

REPAIR

The REPAIR SQL keyword.

REPEATABLE

The REPEATABLE SQL keyword.

REPLACE

The REPLACE SQL keyword.

REPLACE_INVALID_CHARACTERS

The REPLACE_INVALID_CHARACTERS SQL keyword.

REPLICA

The REPLICA SQL keyword.

REPLICATE

The REPLICATE SQL keyword.

REPLICATION

The REPLICATION SQL keyword.

REQUIRE

The REQUIRE SQL keyword.

RESERVED_FOR_COLUMN_ALIAS

Can’t be used as a column alias, so that SELECT <expr> alias can be parsed unambiguously without looking ahead.

RESERVED_FOR_IDENTIFIER

Global list of reserved keywords that cannot be parsed as identifiers without special handling like quoting. Parser should call Dialect::is_reserved_for_identifier to allow for each dialect to customize the list.

RESERVED_FOR_TABLE_ALIAS

These keywords can’t be used as a table alias, so that FROM table_name alias can be parsed unambiguously without looking ahead.

RESERVED_FOR_TABLE_FACTOR

Global list of reserved keywords allowed after FROM. Parser should call Dialect::get_reserved_keyword_after_from to allow for each dialect to customize the list.

RESET

The RESET SQL keyword.

RESOLVE

The RESOLVE SQL keyword.

RESOURCE

The RESOURCE SQL keyword.

RESPECT

The RESPECT SQL keyword.

RESTART

The RESTART SQL keyword.

RESTRICT

The RESTRICT SQL keyword.

RESTRICTED

The RESTRICTED SQL keyword.

RESTRICTIONS

The RESTRICTIONS SQL keyword.

RESTRICTIVE

The RESTRICTIVE SQL keyword.

RESULT

The RESULT SQL keyword.

RESULTSET

The RESULTSET SQL keyword.

RESUME

The RESUME SQL keyword.

RETAIN

The RETAIN SQL keyword.

RETURN

The RETURN SQL keyword.

RETURNING

The RETURNING SQL keyword.

RETURNS

The RETURNS SQL keyword.

REVOKE

The REVOKE SQL keyword.

RIGHT

The RIGHT SQL keyword.

RIGHTARG

The RIGHTARG SQL keyword.

RLIKE

The RLIKE SQL keyword.

RM

The RM SQL keyword.

ROLE

The ROLE SQL keyword.

ROLES

The ROLES SQL keyword.

ROLLBACK

The ROLLBACK SQL keyword.

ROLLUP

The ROLLUP SQL keyword.

ROOT

The ROOT SQL keyword.

ROW

The ROW SQL keyword.

ROWGROUPSIZE

The ROWGROUPSIZE SQL keyword.

ROWID

The ROWID SQL keyword.

ROWS

The ROWS SQL keyword.

ROW_FORMAT

The ROW_FORMAT SQL keyword.

ROW_NUMBER

The ROW_NUMBER SQL keyword.

RULE

The RULE SQL keyword.

RUN

The RUN SQL keyword.

SAFE

The SAFE SQL keyword.

SAFE_CAST

The SAFE_CAST SQL keyword.

SAMPLE

The SAMPLE SQL keyword.

SAVEPOINT

The SAVEPOINT SQL keyword.

SCHEMA

The SCHEMA SQL keyword.

SCHEMAS

The SCHEMAS SQL keyword.

SCOPE

The SCOPE SQL keyword.

SCROLL

The SCROLL SQL keyword.

SEARCH

The SEARCH SQL keyword.

SECOND

The SECOND SQL keyword.

SECONDARY

The SECONDARY SQL keyword.

SECONDARY_ENGINE_ATTRIBUTE

The SECONDARY_ENGINE_ATTRIBUTE SQL keyword.

SECONDS

The SECONDS SQL keyword.

SECRET

The SECRET SQL keyword.

SECURE

The SECURE SQL keyword.

SECURITY

The SECURITY SQL keyword.

SEED

The SEED SQL keyword.

SELECT

The SELECT SQL keyword.

SEMANTIC_VIEW

The SEMANTIC_VIEW SQL keyword.

SEMI

The SEMI SQL keyword.

SEND

The SEND SQL keyword.

SENSITIVE

The SENSITIVE SQL keyword.

SEPARATOR

The SEPARATOR SQL keyword.

SEQUENCE

The SEQUENCE SQL keyword.

SEQUENCEFILE

The SEQUENCEFILE SQL keyword.

SEQUENCES

The SEQUENCES SQL keyword.

SERDE

The SERDE SQL keyword.

SERDEPROPERTIES

The SERDEPROPERTIES SQL keyword.

SERIALIZABLE

The SERIALIZABLE SQL keyword.

SERVER

The SERVER SQL keyword.

SERVICE

The SERVICE SQL keyword.

SESSION

The SESSION SQL keyword.

SESSION_USER

The SESSION_USER SQL keyword.

SET

The SET SQL keyword.

SETERROR

The SETERROR SQL keyword.

SETS

The SETS SQL keyword.

SETTINGS

The SETTINGS SQL keyword.

SHARE

The SHARE SQL keyword.

SHARED

The SHARED SQL keyword.

SHARING

The SHARING SQL keyword.

SHOW

The SHOW SQL keyword.

SIGNED

The SIGNED SQL keyword.

SIMILAR

The SIMILAR SQL keyword.

SIMPLE

The SIMPLE SQL keyword.

SIZE

The SIZE SQL keyword.

SKIP

The SKIP SQL keyword.

SLOW

The SLOW SQL keyword.

SMALLINT

The SMALLINT SQL keyword.

SNAPSHOT

The SNAPSHOT SQL keyword.

SOME

The SOME SQL keyword.

SORT

The SORT SQL keyword.

SORTED

The SORTED SQL keyword.

SOURCE

The SOURCE SQL keyword.

SPATIAL

The SPATIAL SQL keyword.

SPECIFIC

The SPECIFIC SQL keyword.

SPECIFICTYPE

The SPECIFICTYPE SQL keyword.

SPGIST

The SPGIST SQL keyword.

SQL

The SQL SQL keyword.

SQLEXCEPTION

The SQLEXCEPTION SQL keyword.

SQLSTATE

The SQLSTATE SQL keyword.

SQLWARNING

The SQLWARNING SQL keyword.

SQL_BIG_RESULT

The SQL_BIG_RESULT SQL keyword.

SQL_BUFFER_RESULT

The SQL_BUFFER_RESULT SQL keyword.

SQL_CALC_FOUND_ROWS

The SQL_CALC_FOUND_ROWS SQL keyword.

SQL_NO_CACHE

The SQL_NO_CACHE SQL keyword.

SQL_SMALL_RESULT

The SQL_SMALL_RESULT SQL keyword.

SQRT

The SQRT SQL keyword.

SRID

The SRID SQL keyword.

STABLE

The STABLE SQL keyword.

STAGE

The STAGE SQL keyword.

START

The START SQL keyword.

STARTS

The STARTS SQL keyword.

STATEMENT

The STATEMENT SQL keyword.

STATIC

The STATIC SQL keyword.

STATISTICS

The STATISTICS SQL keyword.

STATS_AUTO_RECALC

The STATS_AUTO_RECALC SQL keyword.

STATS_PERSISTENT

The STATS_PERSISTENT SQL keyword.

STATS_SAMPLE_PAGES

The STATS_SAMPLE_PAGES SQL keyword.

STATUPDATE

The STATUPDATE SQL keyword.

STATUS

The STATUS SQL keyword.

STDDEV_POP

The STDDEV_POP SQL keyword.

STDDEV_SAMP

The STDDEV_SAMP SQL keyword.

STDIN

The STDIN SQL keyword.

STDOUT

The STDOUT SQL keyword.

STEP

The STEP SQL keyword.

STORAGE

The STORAGE SQL keyword.

STORAGE_INTEGRATION

The STORAGE_INTEGRATION SQL keyword.

STORAGE_SERIALIZATION_POLICY

The STORAGE_SERIALIZATION_POLICY SQL keyword.

STORED

The STORED SQL keyword.

STRAIGHT_JOIN

The STRAIGHT_JOIN SQL keyword.

STREAM

The STREAM SQL keyword.

STRICT

The STRICT SQL keyword.

STRING

The STRING SQL keyword.

STRUCT

The STRUCT SQL keyword.

SUBMULTISET

The SUBMULTISET SQL keyword.

SUBSCRIPT

The SUBSCRIPT SQL keyword.

SUBSTR

The SUBSTR SQL keyword.

SUBSTRING

The SUBSTRING SQL keyword.

SUBSTRING_REGEX

The SUBSTRING_REGEX SQL keyword.

SUBTYPE

The SUBTYPE SQL keyword.

SUBTYPE_DIFF

The SUBTYPE_DIFF SQL keyword.

SUBTYPE_OPCLASS

The SUBTYPE_OPCLASS SQL keyword.

SUCCEEDS

The SUCCEEDS SQL keyword.

SUM

The SUM SQL keyword.

SUPER

The SUPER SQL keyword.

SUPERUSER

The SUPERUSER SQL keyword.

SUPPORT

The SUPPORT SQL keyword.

SUSPEND

The SUSPEND SQL keyword.

SWAP

The SWAP SQL keyword.

SYMMETRIC

The SYMMETRIC SQL keyword.

SYNC

The SYNC SQL keyword.

SYNONYM

The SYNONYM SQL keyword.

SYSTEM

The SYSTEM SQL keyword.

SYSTEM_TIME

The SYSTEM_TIME SQL keyword.

SYSTEM_USER

The SYSTEM_USER SQL keyword.

TABLE

The TABLE SQL keyword.

TABLES

The TABLES SQL keyword.

TABLESAMPLE

The TABLESAMPLE SQL keyword.

TABLESPACE

The TABLESPACE SQL keyword.

TAG

The TAG SQL keyword.

TARGET

The TARGET SQL keyword.

TARGET_LAG

The TARGET_LAG SQL keyword.

TASK

The TASK SQL keyword.

TBLPROPERTIES

The TBLPROPERTIES SQL keyword.

TEMP

The TEMP SQL keyword.

TEMPORARY

The TEMPORARY SQL keyword.

TEMPTABLE

The TEMPTABLE SQL keyword.

TERMINATED

The TERMINATED SQL keyword.

TERSE

The TERSE SQL keyword.

TEXT

The TEXT SQL keyword.

TEXTFILE

The TEXTFILE SQL keyword.

THEN

The THEN SQL keyword.

TIES

The TIES SQL keyword.

TIME

The TIME SQL keyword.

TIMEFORMAT

The TIMEFORMAT SQL keyword.

TIMESTAMP

The TIMESTAMP SQL keyword.

TIMESTAMPTZ

The TIMESTAMPTZ SQL keyword.

TIMESTAMP_NTZ

The TIMESTAMP_NTZ SQL keyword.

TIMETZ

The TIMETZ SQL keyword.

TIMEZONE

The TIMEZONE SQL keyword.

TIMEZONE_ABBR

The TIMEZONE_ABBR SQL keyword.

TIMEZONE_HOUR

The TIMEZONE_HOUR SQL keyword.

TIMEZONE_MINUTE

The TIMEZONE_MINUTE SQL keyword.

TIMEZONE_REGION

The TIMEZONE_REGION SQL keyword.

TINYBLOB

The TINYBLOB SQL keyword.

TINYINT

The TINYINT SQL keyword.

TINYTEXT

The TINYTEXT SQL keyword.

TO

The TO SQL keyword.

TOP

The TOP SQL keyword.

TOTALS

The TOTALS SQL keyword.

TOTP

The TOTP SQL keyword.

TRACE

The TRACE SQL keyword.

TRAILING

The TRAILING SQL keyword.

TRANSACTION

The TRANSACTION SQL keyword.

TRANSIENT

The TRANSIENT SQL keyword.

TRANSLATE

The TRANSLATE SQL keyword.

TRANSLATE_REGEX

The TRANSLATE_REGEX SQL keyword.

TRANSLATION

The TRANSLATION SQL keyword.

TREAT

The TREAT SQL keyword.

TREE

The TREE SQL keyword.

TRIGGER

The TRIGGER SQL keyword.

TRIM

The TRIM SQL keyword.

TRIM_ARRAY

The TRIM_ARRAY SQL keyword.

TRUE

The TRUE SQL keyword.

TRUNCATE

The TRUNCATE SQL keyword.

TRUNCATECOLUMNS

The TRUNCATECOLUMNS SQL keyword.

TRY

The TRY SQL keyword.

TRY_CAST

The TRY_CAST SQL keyword.

TRY_CONVERT

The TRY_CONVERT SQL keyword.

TSQUERY

The TSQUERY SQL keyword.

TSVECTOR

The TSVECTOR SQL keyword.

TUPLE

The TUPLE SQL keyword.

TYPE

The TYPE SQL keyword.

TYPMOD_IN

The TYPMOD_IN SQL keyword.

TYPMOD_OUT

The TYPMOD_OUT SQL keyword.

UBIGINT

The UBIGINT SQL keyword.

UESCAPE

The UESCAPE SQL keyword.

UHUGEINT

The UHUGEINT SQL keyword.

UINT8

The UINT8 SQL keyword.

UINT16

The UINT16 SQL keyword.

UINT32

The UINT32 SQL keyword.

UINT64

The UINT64 SQL keyword.

UINT128

The UINT128 SQL keyword.

UINT256

The UINT256 SQL keyword.

UNBOUNDED

The UNBOUNDED SQL keyword.

UNCACHE

The UNCACHE SQL keyword.

UNCOMMITTED

The UNCOMMITTED SQL keyword.

UNDEFINED

The UNDEFINED SQL keyword.

UNFREEZE

The UNFREEZE SQL keyword.

UNION

The UNION SQL keyword.

UNIQUE

The UNIQUE SQL keyword.

UNKNOWN

The UNKNOWN SQL keyword.

UNLISTEN

The UNLISTEN SQL keyword.

UNLOAD

The UNLOAD SQL keyword.

UNLOCK

The UNLOCK SQL keyword.

UNLOGGED

The UNLOGGED SQL keyword.

UNMATCHED

The UNMATCHED SQL keyword.

UNNEST

The UNNEST SQL keyword.

UNPIVOT

The UNPIVOT SQL keyword.

UNSAFE

The UNSAFE SQL keyword.

UNSET

The UNSET SQL keyword.

UNSIGNED

The UNSIGNED SQL keyword.

UNTIL

The UNTIL SQL keyword.

UPDATE

The UPDATE SQL keyword.

UPPER

The UPPER SQL keyword.

URL

The URL SQL keyword.

USAGE

The USAGE SQL keyword.

USE

The USE SQL keyword.

USER

The USER SQL keyword.

USER_RESOURCES

The USER_RESOURCES SQL keyword.

USING

The USING SQL keyword.

USMALLINT

The USMALLINT SQL keyword.

UTINYINT

The UTINYINT SQL keyword.

UUID

The UUID SQL keyword.

VACUUM

The VACUUM SQL keyword.

VALID

The VALID SQL keyword.

VALIDATE

The VALIDATE SQL keyword.

VALIDATION_MODE

The VALIDATION_MODE SQL keyword.

VALUE

The VALUE SQL keyword.

VALUES

The VALUES SQL keyword.

VALUE_OF

The VALUE_OF SQL keyword.

VARBINARY

The VARBINARY SQL keyword.

VARBIT

The VARBIT SQL keyword.

VARCHAR

The VARCHAR SQL keyword.

VARCHAR2

The VARCHAR2 SQL keyword.

VARIABLE

The VARIABLE SQL keyword.

VARIABLES

The VARIABLES SQL keyword.

VARYING

The VARYING SQL keyword.

VAR_POP

The VAR_POP SQL keyword.

VAR_SAMP

The VAR_SAMP SQL keyword.

VERBOSE

The VERBOSE SQL keyword.

VERSION

The VERSION SQL keyword.

VERSIONING

The VERSIONING SQL keyword.

VERSIONS

The VERSIONS SQL keyword.

VIEW

The VIEW SQL keyword.

VIEWS

The VIEWS SQL keyword.

VIRTUAL

The VIRTUAL SQL keyword.

VOLATILE

The VOLATILE SQL keyword.

VOLUME

The VOLUME SQL keyword.

WAREHOUSE

The WAREHOUSE SQL keyword.

WAREHOUSES

The WAREHOUSES SQL keyword.

WEEK

The WEEK SQL keyword.

WEEKS

The WEEKS SQL keyword.

WHEN

The WHEN SQL keyword.

WHENEVER

The WHENEVER SQL keyword.

WHERE

The WHERE SQL keyword.

WHILE

The WHILE SQL keyword.

WIDTH_BUCKET

The WIDTH_BUCKET SQL keyword.

WINDOW

The WINDOW SQL keyword.

WITH

The WITH SQL keyword.

WITHIN

The WITHIN SQL keyword.

WITHOUT

The WITHOUT SQL keyword.

WITHOUT_ARRAY_WRAPPER

The WITHOUT_ARRAY_WRAPPER SQL keyword.

WORK

The WORK SQL keyword.

WORKLOAD_IDENTITY

The WORKLOAD_IDENTITY SQL keyword.

WRAPPER

The WRAPPER SQL keyword.

WRITE

The WRITE SQL keyword.

XML

The XML SQL keyword.

XMLNAMESPACES

The XMLNAMESPACES SQL keyword.

XMLTABLE

The XMLTABLE SQL keyword.

XOR

The XOR SQL keyword.

YEAR

The YEAR SQL keyword.

YEARS

The YEARS SQL keyword.

ZONE

The ZONE SQL keyword.

ZORDER

The ZORDER SQL keyword.

ZSTD

The ZSTD SQL keyword.