sqlparser

Module tokenizer

SQL Tokenizer

The tokenizer (a.k.a. lexer) converts a string into a sequence of tokens.

The tokens then form the input for the parser, which outputs an Abstract Syntax Tree (AST).

Structs

Location

Location in input string

Span

A span represents a linear portion of the input string (start, end)

TokenWithSpan

A Token with Span attached to it

Tokenizer

SQL Tokenizer

TokenizerError

Tokenizer error

Word

A keyword (like SELECT) or an optionally quoted SQL identifier

Enums

Token

SQL Token enumeration

Whitespace

Type Aliases

TokenWithLocationDeprecated

Backwards compatibility struct for TokenWithSpan