sqlparser::tokenizer

Enum Token

pub enum Token {

    EOF,
    Word(Word),
    Number(String, bool),
    Char(char),
    SingleQuotedString(String),
    DoubleQuotedString(String),
    TripleSingleQuotedString(String),
    TripleDoubleQuotedString(String),
    DollarQuotedString(DollarQuotedString),
    SingleQuotedByteStringLiteral(String),
    DoubleQuotedByteStringLiteral(String),
    TripleSingleQuotedByteStringLiteral(String),
    TripleDoubleQuotedByteStringLiteral(String),
    SingleQuotedRawStringLiteral(String),
    DoubleQuotedRawStringLiteral(String),
    TripleSingleQuotedRawStringLiteral(String),
    TripleDoubleQuotedRawStringLiteral(String),
    NationalStringLiteral(String),
    EscapedStringLiteral(String),
    UnicodeStringLiteral(String),
    HexStringLiteral(String),
    Comma,
    Whitespace(Whitespace),
    DoubleEq,
    Eq,
    Neq,
    Lt,
    Gt,
    LtEq,
    GtEq,
    Spaceship,
    Plus,
    Minus,
    Mul,
    Div,
    DuckIntDiv,
    Mod,
    StringConcat,
    LParen,
    RParen,
    Period,
    Colon,
    DoubleColon,
    Assignment,
    SemiColon,
    Backslash,
    LBracket,
    RBracket,
    Ampersand,
    Pipe,
    Caret,
    LBrace,
    RBrace,
    RArrow,
    Sharp,
    DoubleSharp,
    Tilde,
    TildeAsterisk,
    ExclamationMarkTilde,
    ExclamationMarkTildeAsterisk,
    DoubleTilde,
    DoubleTildeAsterisk,
    ExclamationMarkDoubleTilde,
    ExclamationMarkDoubleTildeAsterisk,
    ShiftLeft,
    ShiftRight,
    Overlap,
    ExclamationMark,
    DoubleExclamationMark,
    AtSign,
    CaretAt,
    PGSquareRoot,
    PGCubeRoot,
    Placeholder(String),
    Arrow,
    LongArrow,
    HashArrow,
    AtDashAt,
    QuestionMarkDash,
    AmpersandLeftAngleBracket,
    AmpersandRightAngleBracket,
    AmpersandLeftAngleBracketVerticalBar,
    VerticalBarAmpersandRightAngleBracket,
    TwoWayArrow,
    LeftAngleBracketCaret,
    RightAngleBracketCaret,
    QuestionMarkSharp,
    QuestionMarkDashVerticalBar,
    QuestionMarkDoubleVerticalBar,
    TildeEqual,
    ShiftLeftVerticalBar,
    VerticalBarShiftRight,
    VerticalBarRightAngleBracket,
    HashLongArrow,
    AtArrow,
    ArrowAt,
    HashMinus,
    AtQuestion,
    AtAt,
    Question,
    QuestionAnd,
    QuestionPipe,
    CustomBinaryOperator(String),

}

SQL Token enumeration

Variants

EOF

An end-of-file marker, not a real token

Word(Word)

A keyword (like SELECT) or an optionally quoted SQL identifier

Number(String, bool)

An unsigned numeric literal

Char(char)

A character that could not be tokenized

SingleQuotedString(String)

Single quoted string: i.e: ‘string’

DoubleQuotedString(String)

Double quoted string: i.e: “string”

TripleSingleQuotedString(String)

Triple single quoted strings: Example ‘’‘abc’‘’ BigQuery

TripleDoubleQuotedString(String)

Triple double quoted strings: Example “”“abc”“” BigQuery

DollarQuotedString(DollarQuotedString)

Dollar quoted string: i.e: $$string$$ or $tag_name$string$tag_name$

SingleQuotedByteStringLiteral(String)

Byte string literal: i.e: b’string’ or B’string’ (note that some backends, such as PostgreSQL, may treat this syntax as a bit string literal instead, i.e: b’10010101’)

DoubleQuotedByteStringLiteral(String)

Byte string literal: i.e: b“string“ or B“string“

TripleSingleQuotedByteStringLiteral(String)

Triple single quoted literal with byte string prefix. Example B'''abc''' BigQuery

TripleDoubleQuotedByteStringLiteral(String)

Triple double quoted literal with byte string prefix. Example B"""abc""" BigQuery

SingleQuotedRawStringLiteral(String)

Single quoted literal with raw string prefix. Example R'abc' BigQuery

DoubleQuotedRawStringLiteral(String)

Double quoted literal with raw string prefix. Example R"abc" BigQuery

TripleSingleQuotedRawStringLiteral(String)

Triple single quoted literal with raw string prefix. Example R'''abc''' BigQuery

TripleDoubleQuotedRawStringLiteral(String)

Triple double quoted literal with raw string prefix. Example R"""abc""" BigQuery

NationalStringLiteral(String)

“National” string literal: i.e: N’string’

EscapedStringLiteral(String)

“escaped” string literal, which are an extension to the SQL standard: i.e: e’first \n second’ or E ‘first \n second’

UnicodeStringLiteral(String)

Unicode string literal: i.e: U&‘first \000A second’

HexStringLiteral(String)

Hexadecimal string literal: i.e.: X’deadbeef’

Comma

Comma

Whitespace(Whitespace)

Whitespace (space, tab, etc)

DoubleEq

Double equals sign ==

Eq

Equality operator =

Neq

Not Equals operator <> (or != in some dialects)

Lt

Less Than operator <

Gt

Greater Than operator >

LtEq

Less Than Or Equals operator <=

GtEq

Greater Than Or Equals operator >=

Spaceship

Spaceship operator <=>

Plus

Plus operator +

Minus

Minus operator -

Mul

Multiplication operator *

Div

Division operator /

DuckIntDiv

Integer division operator // in DuckDB

Mod

Modulo Operator %

StringConcat

String concatenation ||

LParen

Left parenthesis (

RParen

Right parenthesis )

Period

Period (used for compound identifiers or projections into nested types)

Colon

Colon :

DoubleColon

DoubleColon :: (used for casting in PostgreSQL)

Assignment

Assignment := (used for keyword argument in DuckDB macros and some functions, and for variable declarations in DuckDB and Snowflake)

SemiColon

SemiColon ; used as separator for COPY and payload

Backslash

Backslash \ used in terminating the COPY payload with \.

LBracket

Left bracket [

RBracket

Right bracket ]

Ampersand

Ampersand &

Pipe

Pipe |

Caret

Caret ^

LBrace

Left brace {

RBrace

Right brace }

RArrow

Right Arrow =>

Sharp

Sharp # used for PostgreSQL Bitwise XOR operator, also PostgreSQL/Redshift geometrical unary/binary operator (Number of points in path or polygon/Intersection)

DoubleSharp

## PostgreSQL/Redshift geometrical binary operator (Point of closest proximity)

Tilde

Tilde ~ used for PostgreSQL Bitwise NOT operator or case sensitive match regular expression operator

TildeAsterisk

~* , a case insensitive match regular expression operator in PostgreSQL

ExclamationMarkTilde

!~ , a case sensitive not match regular expression operator in PostgreSQL

ExclamationMarkTildeAsterisk

!~* , a case insensitive not match regular expression operator in PostgreSQL

DoubleTilde

~~, a case sensitive match pattern operator in PostgreSQL

DoubleTildeAsterisk

~~*, a case insensitive match pattern operator in PostgreSQL

ExclamationMarkDoubleTilde

!~~, a case sensitive not match pattern operator in PostgreSQL

ExclamationMarkDoubleTildeAsterisk

!~~*, a case insensitive not match pattern operator in PostgreSQL

ShiftLeft

<<, a bitwise shift left operator in PostgreSQL

ShiftRight

>>, a bitwise shift right operator in PostgreSQL

Overlap

&&, an overlap operator in PostgreSQL

ExclamationMark

Exclamation Mark ! used for PostgreSQL factorial operator

DoubleExclamationMark

Double Exclamation Mark !! used for PostgreSQL prefix factorial operator

AtSign

AtSign @ used for PostgreSQL abs operator, also PostgreSQL/Redshift geometrical unary/binary operator (Center, Contained or on)

CaretAt

^@, a “starts with” string operator in PostgreSQL

PGSquareRoot

|/, a square root math operator in PostgreSQL

PGCubeRoot

||/, a cube root math operator in PostgreSQL

Placeholder(String)

? or $ , a prepared statement arg placeholder

Arrow

->, used as a operator to extract json field in PostgreSQL

LongArrow

->>, used as a operator to extract json field as text in PostgreSQL

HashArrow

#>, extracts JSON sub-object at the specified path

AtDashAt

@-@ PostgreSQL/Redshift geometrical unary operator (Length or circumference)

QuestionMarkDash

?- PostgreSQL/Redshift geometrical unary/binary operator (Is horizontal?/Are horizontally aligned?)

AmpersandLeftAngleBracket

&< PostgreSQL/Redshift geometrical binary operator (Overlaps to left?)

AmpersandRightAngleBracket

&> PostgreSQL/Redshift geometrical binary operator (Overlaps to right?)`

AmpersandLeftAngleBracketVerticalBar

&<| PostgreSQL/Redshift geometrical binary operator (Does not extend above?)`

VerticalBarAmpersandRightAngleBracket

|&> PostgreSQL/Redshift geometrical binary operator (Does not extend below?)`

TwoWayArrow

<-> PostgreSQL/Redshift geometrical binary operator (Distance between)

LeftAngleBracketCaret

<^ PostgreSQL/Redshift geometrical binary operator (Is below?)

RightAngleBracketCaret

>^ PostgreSQL/Redshift geometrical binary operator (Is above?)

QuestionMarkSharp

?# PostgreSQL/Redshift geometrical binary operator (Intersects or overlaps)

QuestionMarkDashVerticalBar

?-| PostgreSQL/Redshift geometrical binary operator (Is perpendicular?)

QuestionMarkDoubleVerticalBar

?|| PostgreSQL/Redshift geometrical binary operator (Are parallel?)

TildeEqual

~= PostgreSQL/Redshift geometrical binary operator (Same as)

ShiftLeftVerticalBar

`<<| PostgreSQL/Redshift geometrical binary operator (Is strictly below?)

VerticalBarShiftRight

`|>> PostgreSQL/Redshift geometrical binary operator (Is strictly above?)

VerticalBarRightAngleBracket

`|> BigQuery pipe operator

HashLongArrow

#>>, extracts JSON sub-object at the specified path as text

AtArrow

jsonb @> jsonb -> boolean: Test whether left json contains the right json

ArrowAt

jsonb <@ jsonb -> boolean: Test whether right json contains the left json

HashMinus

jsonb #- text[] -> jsonb: Deletes the field or array element at the specified path, where path elements can be either field keys or array indexes.

AtQuestion

jsonb @? jsonpath -> boolean: Does JSON path return any item for the specified JSON value?

AtAt

jsonb @@ jsonpath → boolean: Returns the result of a JSON path predicate check for the specified JSON value. Only the first item of the result is taken into account. If the result is not Boolean, then NULL is returned.

Question

jsonb ? text -> boolean: Checks whether the string exists as a top-level key within the jsonb object

QuestionAnd

jsonb ?& text[] -> boolean: Check whether all members of the text array exist as top-level keys within the jsonb object

QuestionPipe

jsonb ?| text[] -> boolean: Check whether any member of the text array exists as top-level keys within the jsonb object

CustomBinaryOperator(String)

Custom binary operator This is used to represent any custom binary operator that is not part of the SQL standard. PostgreSQL allows defining custom binary operators using CREATE OPERATOR.

Implementations

impl Token

pub fn make_keyword(keyword: &str) -> Self

pub fn make_word(word: &str, quote_style: Option<char>) -> Self