Module sqlparser::ast

SQL Abstract Syntax Tree (AST) types

Modules

helpers

Structs

Array

Represents an Array Expression, either ARRAY[..], or [..]

ArrayAgg

An ARRAY_AGG invocation ARRAY_AGG( [ DISTINCT ] <expr> [ORDER BY <expr>] [LIMIT <n>] ) Or ARRAY_AGG( [ DISTINCT ] <expr> ) [ WITHIN GROUP ( ORDER BY <expr> ) ] ORDER BY position is defined differently for BigQuery, Postgres and Snowflake.

Assignment

SQL assignment foo = expr as used in SQLUpdate

CharacterLength

Information about character length, including length and possibly unit.

ColumnDef

SQL column definition

ColumnOptionDef

An optionally-named ColumnOption: [ CONSTRAINT <name> ] <column-option>.

CreateFunctionBody

Postgres specific feature.

Cte

A single CTE (used after WITH): alias [(col1, col2, ...)] AS ( query ) The names in the column list before AS, when specified, replace the names of the columns returned by the query. The parser does not validate that the number of columns in the query matches the number of columns in the query.

DoUpdate

DollarQuotedString

DropFunctionDesc

Function describe in DROP FUNCTION.

ExceptSelectItem

Bigquery EXCEPT information, with at least one column.

Fetch

Function

A function call

HiveFormat

Ident

An identifier, decomposed into its value or character data and the quote style.

IdentWithAlias

Single aliased identifier

Join

LateralView

A hive LATERAL VIEW with potential column aliases

ListAgg

A LISTAGG invocation LISTAGG( [ DISTINCT ] <expr>[, <separator> ] [ON OVERFLOW <on_overflow>] ) ) [ WITHIN GROUP (ORDER BY <within_group1>[, ...] ) ]

LockClause

ObjectName

A name of a table, view, custom type, etc., possibly multi-part, i.e. db.schema.obj

Offset

OnConflict

OperateFunctionArg

Function argument in CREATE OR DROP FUNCTION.

OrderByExpr

An ORDER BY expression

Query

The most complete variant of a SELECT query expression, optionally including WITH, UNION / other set operations, and ORDER BY.

ReplaceSelectElement

Syntax

ReplaceSelectItem

Bigquery REPLACE information.

Select

A restricted variant of SELECT (without CTEs/ORDER BY), which may appear either as the only body item of a Query, or as an operand to a set operation like UNION.

SelectInto

SqlOption

Table

A TABLE command

TableAlias

TableWithJoins

Top

Values

WildcardAdditionalOptions

Additional options for wildcards, e.g. Snowflake EXCLUDE/RENAME and Bigquery EXCEPT.

WindowFrame

Specifies the data processed by a window function, e.g. RANGE UNBOUNDED PRECEDING or ROWS BETWEEN 5 PRECEDING AND CURRENT ROW.

WindowSpec

A window specification (i.e. OVER (PARTITION BY .. ORDER BY .. etc.))

With

Enums

Action

A privilege on a database object (table, sequence, etc.).

AddDropSync

AlterColumnOperation

An ALTER COLUMN (Statement::AlterTable) operation

AlterIndexOperation

AlterTableOperation

An ALTER TABLE (Statement::AlterTable) operation

AnalyzeFormat

ArgMode

The mode of an argument in CREATE FUNCTION.

BinaryOperator

Binary operators

CharLengthUnits

Possible units for characters, initially based on 2016 ANSI standard.

CloseCursor

ColumnOption

ColumnOptions are modifiers that follow a column definition in a CREATE TABLE statement.

CommentObject

ConflictTarget

ContextModifier

Optional context modifier for statements that can be or LOCAL, or SESSION.

CopyLegacyCsvOption

A CSV option in COPY statement before PostgreSQL version 9.0.

CopyLegacyOption

An option in COPY statement before PostgreSQL version 9.0.

CopyOption

An option in COPY statement.

CopyTarget

CreateFunctionUsing

DataType

SQL data types

DateTimeField

DiscardObject

DropFunctionOption

Function describe in DROP FUNCTION.

ExactNumberInfo

Additional information for NUMERIC, DECIMAL, and DEC data types following the 2016 standard.

ExcludeSelectItem

Snowflake EXCLUDE information.

Expr

An SQL expression of any type.

FetchDirection

Specific direction for FETCH statement

FileFormat

External table’s available file format

FunctionArg

FunctionArgExpr

FunctionBehavior

These attributes inform the query optimizer about the behavior of the function.

FunctionDefinition

GrantObjects

Objects on which privileges are granted in a GRANT statement.

HiveDistributionStyle

HiveIOFormat

HiveRowFormat

IndexType

Indexing method used by that index.

JoinConstraint

JoinOperator

JsonOperator

JsonOperator

KeyOrIndexDisplay

Representation whether a definition can can contains the KEY or INDEX keywords with the same meaning.

KillType

ListAggOnOverflow

The ON OVERFLOW clause of a LISTAGG invocation

LockType

MergeClause

MinMaxValue

Can use to describe options in create sequence or table column type identity [ MINVALUE minvalue | NO MINVALUE ] [ MAXVALUE maxvalue | NO MAXVALUE ]

NonBlock

ObjectType

OffsetRows

Stores the keyword after OFFSET <number>

OnCommit

OnConflictAction

OnInsert

Password

Privileges

Privileges granted in a GRANT statement or revoked in a REVOKE statement.

ReferentialAction

<referential_action> = { RESTRICT | CASCADE | SET NULL | NO ACTION | SET DEFAULT }

RenameSelectItem

Snowflake RENAME information.

SchemaName

Schema possible naming variants (1).

SearchModifier

Fulltext search modifiers (1).

SelectItem

One item of the comma-separated list following SELECT

SequenceOptions

Can use to describe options in create sequence or table column type identity

SetExpr

A node in a tree, representing a “query body” expression, roughly: SELECT ... [ {UNION|EXCEPT|INTERSECT} SELECT ...]

SetOperator

SetQuantifier

A quantifier for SetOperator.

ShowCreateObject

ShowStatementFilter

SqliteOnConflict

Sqlite specific syntax

Statement

A top-level statement (SELECT, INSERT, CREATE, etc.)

TableConstraint

A table-level constraint, specified in a CREATE TABLE or an ALTER TABLE ADD <constraint> statement.

TableFactor

A table name or a parenthesized subquery with an optional alias

TimezoneInfo

Timestamp and Time data types information about TimeZone formatting.

TransactionAccessMode

TransactionIsolationLevel

TransactionMode

TrimWhereField

UnaryOperator

Unary operators

Value

Primitive SQL values such as number and string

WindowFrameBound

Specifies WindowFrame’s start_bound and end_bound

WindowFrameUnits

Functions

escape_quoted_string