Enum sqlparser::ast::DataType

pub enum DataType {

    Character(Option<CharacterLength>),
    Char(Option<CharacterLength>),
    CharacterVarying(Option<CharacterLength>),
    CharVarying(Option<CharacterLength>),
    Varchar(Option<CharacterLength>),
    Nvarchar(Option<u64>),
    Uuid,
    CharacterLargeObject(Option<u64>),
    CharLargeObject(Option<u64>),
    Clob(Option<u64>),
    Binary(Option<u64>),
    Varbinary(Option<u64>),
    Blob(Option<u64>),
    Numeric(ExactNumberInfo),
    Decimal(ExactNumberInfo),
    BigNumeric(ExactNumberInfo),
    BigDecimal(ExactNumberInfo),
    Dec(ExactNumberInfo),
    Float(Option<u64>),
    TinyInt(Option<u64>),
    UnsignedTinyInt(Option<u64>),
    SmallInt(Option<u64>),
    UnsignedSmallInt(Option<u64>),
    MediumInt(Option<u64>),
    UnsignedMediumInt(Option<u64>),
    Int(Option<u64>),
    Integer(Option<u64>),
    UnsignedInt(Option<u64>),
    UnsignedInteger(Option<u64>),
    BigInt(Option<u64>),
    UnsignedBigInt(Option<u64>),
    Real,
    Double,
    DoublePrecision,
    Boolean,
    Date,
    Time(Option<u64>, TimezoneInfo),
    Datetime(Option<u64>),
    Timestamp(Option<u64>, TimezoneInfo),
    Interval,
    JSON,
    Regclass,
    Text,
    String,
    Bytea,
    Custom(ObjectName, Vec<String>),
    Array(Option<Box<DataType>>),
    Enum(Vec<String>),
    Set(Vec<String>),

}

SQL data types

Variants

Character(Option<CharacterLength>)

Fixed-length character type e.g. CHARACTER(10)

Char(Option<CharacterLength>)

Fixed-length char type e.g. CHAR(10)

CharacterVarying(Option<CharacterLength>)

Character varying type e.g. CHARACTER VARYING(10)

CharVarying(Option<CharacterLength>)

Char varying type e.g. CHAR VARYING(10)

Varchar(Option<CharacterLength>)

Variable-length character type e.g. VARCHAR(10)

Nvarchar(Option<u64>)

Variable-length character type e.g. NVARCHAR(10)

Uuid

Uuid type

CharacterLargeObject(Option<u64>)

Large character object with optional length e.g. CHARACTER LARGE OBJECT, CHARACTER LARGE OBJECT(1000), standard

CharLargeObject(Option<u64>)

Large character object with optional length e.g. CHAR LARGE OBJECT, CHAR LARGE OBJECT(1000), standard

Clob(Option<u64>)

Large character object with optional length e.g. CLOB, CLOB(1000), standard

Binary(Option<u64>)

Fixed-length binary type with optional length e.g. standard, MS SQL Server

Varbinary(Option<u64>)

Variable-length binary with optional length type e.g. standard, MS SQL Server

Blob(Option<u64>)

Large binary object with optional length e.g. BLOB, BLOB(1000), standard, Oracle

Numeric(ExactNumberInfo)

Numeric type with optional precision and scale e.g. NUMERIC(10,2), standard

Decimal(ExactNumberInfo)

Decimal type with optional precision and scale e.g. DECIMAL(10,2), standard

BigNumeric(ExactNumberInfo)

BigNumeric type used in BigQuery

BigDecimal(ExactNumberInfo)

This is alias for BigNumeric type used in BigQuery

Dec(ExactNumberInfo)

Dec type with optional precision and scale e.g. DEC(10,2), standard

Float(Option<u64>)

Floating point with optional precision e.g. FLOAT(8)

TinyInt(Option<u64>)

Tiny integer with optional display width e.g. TINYINT or TINYINT(3)

UnsignedTinyInt(Option<u64>)

Unsigned tiny integer with optional display width e.g. TINYINT UNSIGNED or TINYINT(3) UNSIGNED

SmallInt(Option<u64>)

Small integer with optional display width e.g. SMALLINT or SMALLINT(5)

UnsignedSmallInt(Option<u64>)

Unsigned small integer with optional display width e.g. SMALLINT UNSIGNED or SMALLINT(5) UNSIGNED

MediumInt(Option<u64>)

MySQL medium integer (1) with optional display width e.g. MEDIUMINT or MEDIUMINT(5)

UnsignedMediumInt(Option<u64>)

Unsigned medium integer (1) with optional display width e.g. MEDIUMINT UNSIGNED or MEDIUMINT(5) UNSIGNED

Int(Option<u64>)

Integer with optional display width e.g. INT or INT(11)

Integer(Option<u64>)

Integer with optional display width e.g. INTEGER or INTEGER(11)

UnsignedInt(Option<u64>)

Unsigned integer with optional display width e.g. INT UNSIGNED or INT(11) UNSIGNED

UnsignedInteger(Option<u64>)

Unsigned integer with optional display width e.g. INTGER UNSIGNED or INTEGER(11) UNSIGNED

BigInt(Option<u64>)

Big integer with optional display width e.g. BIGINT or BIGINT(20)

UnsignedBigInt(Option<u64>)

Unsigned big integer with optional display width e.g. BIGINT UNSIGNED or BIGINT(20) UNSIGNED

Real

Floating point e.g. REAL

Double

Double

DoublePrecision

Double PRECISION e.g. standard, postgresql

Boolean

Boolean

Date

Date

Time(Option<u64>, TimezoneInfo)

Time with optional time precision and time zone information e.g. standard.

Datetime(Option<u64>)

Datetime with optional time precision e.g. MySQL.

Timestamp(Option<u64>, TimezoneInfo)

Timestamp with optional time precision and time zone information e.g. standard.

Interval

Interval

JSON

JSON type used in BigQuery

Regclass

Regclass used in postgresql serial

Text

Text

String

String

Bytea

Bytea

Custom(ObjectName, Vec<String>)

Custom type such as enums

Array(Option<Box<DataType>>)

Arrays

Enum(Vec<String>)

Enums

Set(Vec<String>)

Set