sqlparser

Module keywords

This module defines:

a list of constants for every keyword

an ALL_KEYWORDS array with every keyword in it This is not a list of reserved keywords: some of these can be parsed as identifiers if the parser decides so. This means that new keywords can be added here without affecting the parse result.

As a matter of fact, most of these keywords are not used at all and could be removed.

a RESERVED_FOR_TABLE_ALIAS array with keywords reserved in a “table alias” context.

Enums

Keyword

Constants

ABORT

ABS

ABSENT

ABSOLUTE

ACCEPTANYDATE

ACCEPTINVCHARS

ACCESS

ACCOUNT

ACTION

ADD

ADDQUOTES

ADMIN

AFTER

AGAINST

AGGREGATE

AGGREGATION

ALERT

ALGORITHM

ALIAS

ALL

ALLOCATE

ALLOWOVERWRITE

ALL_KEYWORDS

ALL_KEYWORDS_INDEX

ALTER

ALWAYS

ANALYZE

AND

ANTI

ANY

APPLICATION

APPLY

APPLYBUDGET

ARCHIVE

ARE

ARRAY

ARRAY_MAX_CARDINALITY

AS

ASC

ASENSITIVE

ASOF

ASSERT

ASYMMETRIC

AT

ATOMIC

ATTACH

AUDIT

AUTHENTICATION

AUTHORIZATION

AUTO

AUTOEXTEND_SIZE

AUTOINCREMENT

AUTO_INCREMENT

AVG

AVG_ROW_LENGTH

AVRO

BACKWARD

BASE64

BASE_LOCATION

BEFORE

BEGIN

BEGIN_FRAME

BEGIN_PARTITION

BERNOULLI

BETWEEN

BIGDECIMAL

BIGINT

BIGNUMERIC

BINARY

BIND

BINDING

BIT

BLANKSASNULL

BLOB

BLOCK

BLOOM

BLOOMFILTER

BOOL

BOOLEAN

BOOST

BOTH

BOX

BRIN

BROWSE

BTREE

BUCKET

BUCKETS

BY

BYPASSRLS

BYTEA

BYTES

BZIP2

CACHE

CALL

CALLED

CARDINALITY

CASCADE

CASCADED

CASE

CASES

CAST

CATALOG

CATALOG_SYNC

CATALOG_SYNC_NAMESPACE_FLATTEN_DELIMITER

CATALOG_SYNC_NAMESPACE_MODE

CATCH

CEIL

CEILING

CENTURY

CHAIN

CHANGE

CHANGE_TRACKING

CHANNEL

CHAR

CHARACTER

CHARACTERISTICS

CHARACTERS

CHARACTER_LENGTH

CHARSET

CHAR_LENGTH

CHECK

CHECKSUM

CIRCLE

CLEANPATH

CLEAR

CLOB

CLONE

CLOSE

CLUSTER

CLUSTERED

CLUSTERING

COALESCE

COLLATE

COLLATION

COLLECT

COLLECTION

COLUMN

COLUMNS

COLUMNSTORE

COMMENT

COMMIT

COMMITTED

COMPATIBLE

COMPRESSION

COMPUTE

CONCURRENTLY

CONDITION

CONFLICT

CONNECT

CONNECTION

CONNECTOR

CONNECT_BY_ROOT

CONSTRAINT

CONTACT

CONTAINS

CONTINUE

CONVERT

COPY

COPY_OPTIONS

CORR

CORRESPONDING

COUNT

COVAR_POP

COVAR_SAMP

CREATE

CREATEDB

CREATEROLE

CREDENTIALS

CROSS

CSV

CUBE

CUME_DIST

CURRENT

CURRENT_CATALOG

CURRENT_DATE

CURRENT_DEFAULT_TRANSFORM_GROUP

CURRENT_PATH

CURRENT_ROLE

CURRENT_ROW

CURRENT_SCHEMA

CURRENT_TIME

CURRENT_TIMESTAMP

CURRENT_TRANSFORM_GROUP_FOR_TYPE

CURRENT_USER

CURSOR

CYCLE

DATA

DATABASE

DATABASES

DATA_RETENTION_TIME_IN_DAYS

DATE

DATE32

DATEFORMAT

DATETIME

DATETIME64

DAY

DAYOFWEEK

DAYOFYEAR

DAYS

DCPROPERTIES

DEALLOCATE

DEC

DECADE

DECIMAL

DECLARE

DEDUPLICATE

DEFAULT

DEFAULTS

DEFAULT_DDL_COLLATION

DEFERRABLE

DEFERRED

DEFINE

DEFINED

DEFINER

DELAYED

DELAY_KEY_WRITE

DELETE

DELIMITED

DELIMITER

DELTA

DENSE_RANK

DENY

DEREF

DESC

DESCRIBE

DETACH

DETAIL

DETERMINISTIC

DIMENSIONS

DIRECTORY

DISABLE

DISCARD

DISCONNECT

DISTINCT

DISTRIBUTE

DIV

DO

DOMAIN

DOUBLE

DOW

DOWNSTREAM

DOY

DROP

DRY

DUPLICATE

DYNAMIC

EACH

ELEMENT

ELEMENTS

ELSE

ELSEIF

EMPTY

EMPTYASNULL

ENABLE

ENABLE_SCHEMA_EVOLUTION

ENCODING

ENCRYPTED

ENCRYPTION

END

ENDPOINT

END_EXEC

END_FRAME

END_PARTITION

ENFORCED

ENGINE

ENGINE_ATTRIBUTE

ENUM

ENUM8

ENUM16

EPHEMERAL

EPOCH

EQUALS

ERROR

ESCAPE

ESCAPED

ESTIMATE

EVENT

EVERY

EVOLVE

EXCEPT

EXCEPTION

EXCHANGE

EXCLUDE

EXCLUDING

EXCLUSIVE

EXEC

EXECUTE

EXECUTION

EXISTS

EXP

EXPANSION

EXPLAIN

EXPLICIT

EXPORT

EXTEND

EXTENDED

EXTENSION

EXTERNAL

EXTERNAL_VOLUME

EXTRACT

FACTS

FAIL

FAILOVER

FALSE

FETCH

FIELDS

FILE

FILES

FILE_FORMAT

FILL

FILTER

FINAL

FIRST

FIRST_VALUE

FIXEDSTRING

FIXEDWIDTH

FLATTEN

FLOAT

FLOAT4

FLOAT8

FLOAT32

FLOAT64

FLOOR

FLUSH

FN

FOLLOWING

FOR

FORCE

FORCE_NOT_NULL

FORCE_NULL

FORCE_QUOTE

FOREIGN

FORMAT

FORMATTED

FORWARD

FRAME_ROW

FREE

FREEZE

FROM

FSCK

FULFILLMENT

FULL

FULLTEXT

FUNCTION

FUNCTIONS

FUSION

FUTURE

GB

GENERAL

GENERATE

GENERATED

GEOGRAPHY

GET

GIN

GIST

GLOBAL

GRANT

GRANTED

GRANTS

GRAPHVIZ

GROUP

GROUPING

GROUPS

GZIP

HASH

HAVING

HEADER

HEAP

HIGH_PRIORITY

HISTORY

HIVEVAR

HOLD

HOSTS

HOUR

HOURS

HUGEINT

IAM_ROLE

ICEBERG

ID

IDENTITY

IDENTITY_INSERT

IF

IGNORE

IGNOREHEADER

ILIKE

IMMEDIATE

IMMUTABLE

IMPORT

IMPORTED

IN

INCLUDE

INCLUDE_NULL_VALUES

INCLUDING

INCREMENT

INCREMENTAL

INDEX

INDICATOR

INHERIT

INHERITS

INITIALIZE

INITIALLY

INNER

INOUT

INPATH

INPLACE

INPUT

INPUTFORMAT

INSENSITIVE

INSERT

INSERT_METHOD

INSTALL

INSTANT

INSTEAD

INT

INT2

INT4

INT8

INT16

INT32

INT64

INT128

INT256

INTEGER

INTEGRATION

INTERPOLATE

INTERSECT

INTERSECTION

INTERVAL

INTO

INVOKER

IO

IS

ISODOW

ISOLATION

ISOWEEK

ISOYEAR

ITEMS

JAR

JOIN

JSON

JSONB

JSONFILE

JSON_TABLE

JULIAN

KEY

KEYS

KEY_BLOCK_SIZE

KILL

LAG

LANGUAGE

LARGE

LAST

LAST_VALUE

LATERAL

LEAD

LEADING

LEFT

LEVEL

LIKE

LIKE_REGEX

LIMIT

LINE

LINES

LIST

LISTEN

LISTING

LN

LOAD

LOCAL

LOCALTIME

LOCALTIMESTAMP

LOCATION

LOCK

LOCKED

LOG

LOGIN

LOGS

LONGBLOB

LONGTEXT

LOWCARDINALITY

LOWER

LOW_PRIORITY

LS

LSEG

MACRO

MANAGE

MANAGED

MANAGEDLOCATION

MANIFEST

MAP

MASKING

MATCH

MATCHED

MATCHES

MATCH_CONDITION

MATCH_RECOGNIZE

MATERIALIZE

MATERIALIZED

MAX

MAXFILESIZE

MAXVALUE

MAX_DATA_EXTENSION_TIME_IN_DAYS

MAX_ROWS

MB

MEASURES

MEDIUMBLOB

MEDIUMINT

MEDIUMTEXT

MEMBER

MERGE

MESSAGE

METADATA

METHOD

METRIC

METRICS

MICROSECOND

MICROSECONDS

MILLENIUM

MILLENNIUM

MILLISECOND

MILLISECONDS

MIN

MINUS

MINUTE

MINUTES

MINVALUE

MIN_ROWS

MOD

MODE

MODIFIES

MODIFY

MODULE

MONITOR

MONTH

MONTHS

MSCK

MULTISET

MUTATION

NAME

NAMES

NANOSECOND

NANOSECONDS

NATIONAL

NATURAL

NCHAR

NCLOB

NEST

NESTED

NETWORK

NEW

NEXT

NFC

NFD

NFKC

NFKD

NO

NOBYPASSRLS

NOCREATEDB

NOCREATEROLE

NOINHERIT

NOLOGIN

NONE

NOORDER

NOREPLICATION

NORMALIZE

NORMALIZED

NOSCAN

NOSUPERUSER

NOT

NOTHING

NOTIFY

NOTNULL

NOWAIT

NO_WRITE_TO_BINLOG

NTH_VALUE

NTILE

NULL

NULLABLE

NULLIF

NULLS

NUMERIC

NVARCHAR

OBJECT

OBJECTS

OCCURRENCES_REGEX

OCTETS

OCTET_LENGTH

OF

OFF

OFFSET

OFFSETS

OLD

OMIT

ON

ONE

ONLY

ON_CREATE

ON_SCHEDULE

OPEN

OPENJSON

OPERATE

OPERATOR

OPTIMIZATION

OPTIMIZE

OPTIMIZED

OPTIMIZER_COSTS

OPTION

OPTIONS

OR

ORC

ORDER

ORDINALITY

ORGANIZATION

OTHER

OUT

OUTER

OUTPUT

OUTPUTFORMAT

OVER

OVERFLOW

OVERLAPS

OVERLAY

OVERRIDE

OVERWRITE

OWNED

OWNER

OWNERSHIP

PACKAGE

PACKAGES

PACK_KEYS

PARALLEL

PARAMETER

PARQUET

PART

PARTITION

PARTITIONED

PARTITIONS

PASSING

PASSWORD

PAST

PATH

PATTERN

PER

PERCENT

PERCENTILE_CONT

PERCENTILE_DISC

PERCENT_RANK

PERIOD

PERMISSIVE

PERSISTENT

PIVOT

PLACING

PLAN

PLANS

POINT

POLICY

POLYGON

POOL

PORTION

POSITION

POSITION_REGEX

POWER

PRAGMA

PRECEDES

PRECEDING

PRECISION

PREPARE

PRESERVE

PREWHERE

PRIMARY

PRINT

PRIOR

PRIVILEGES

PROCEDURE

PROFILE

PROGRAM

PROJECTION

PUBLIC

PURCHASE

PURGE

QUALIFY

QUARTER

QUERY

QUOTE

RAISE

RAISERROR

RANGE

RANK

RAW

RCFILE

READ

READS

READ_ONLY

REAL

RECLUSTER

RECURSIVE

REF

REFERENCES

REFERENCING

REFRESH_MODE

REGCLASS

REGEXP

REGION

REGR_AVGX

REGR_AVGY

REGR_COUNT

REGR_INTERCEPT

REGR_R2

REGR_SLOPE

REGR_SXX

REGR_SXY

REGR_SYY

REINDEX

RELATIVE

RELAY

RELEASE

RELEASES

REMOTE

REMOVE

RENAME

REORG

REPAIR

REPEATABLE

REPLACE

REPLACE_INVALID_CHARACTERS

REPLICA

REPLICATE

REPLICATION

REQUIRE

RESERVED_FOR_COLUMN_ALIAS

Can’t be used as a column alias, so that SELECT <expr> alias can be parsed unambiguously without looking ahead.

RESERVED_FOR_IDENTIFIER

Global list of reserved keywords that cannot be parsed as identifiers without special handling like quoting. Parser should call Dialect::is_reserved_for_identifier to allow for each dialect to customize the list.

RESERVED_FOR_TABLE_ALIAS

These keywords can’t be used as a table alias, so that FROM table_name alias can be parsed unambiguously without looking ahead.

RESERVED_FOR_TABLE_FACTOR

RESET

RESOLVE

RESOURCE

RESPECT

RESTART

RESTRICT

RESTRICTED

RESTRICTIONS

RESTRICTIVE

RESULT

RESULTSET

RESUME

RETAIN

RETURN

RETURNING

RETURNS

REVOKE

RIGHT

RLIKE

RM

ROLE

ROLES

ROLLBACK

ROLLUP

ROOT

ROW

ROWGROUPSIZE

ROWID

ROWS

ROW_FORMAT

ROW_NUMBER

RULE

RUN

SAFE

SAFE_CAST

SAMPLE

SAVEPOINT

SCHEMA

SCHEMAS

SCOPE

SCROLL

SEARCH

SECOND

SECONDARY

SECONDARY_ENGINE_ATTRIBUTE

SECONDS

SECRET

SECURE

SECURITY

SEED

SELECT

SEMANTIC_VIEW

SEMI

SENSITIVE

SEPARATOR

SEQUENCE

SEQUENCEFILE

SEQUENCES

SERDE

SERDEPROPERTIES

SERIALIZABLE

SERVER

SERVICE

SESSION

SESSION_USER

SET

SETERROR

SETS

SETTINGS

SHARE

SHARED

SHARING

SHOW

SIGNED

SIMILAR

SKIP

SLOW

SMALLINT

SNAPSHOT

SOME

SORT

SORTED

SOURCE

SPATIAL

SPECIFIC

SPECIFICTYPE

SPGIST

SQL

SQLEXCEPTION

SQLSTATE

SQLWARNING

SQRT

SRID

STABLE

STAGE

START

STARTS

STATEMENT

STATIC

STATISTICS

STATS_AUTO_RECALC

STATS_PERSISTENT

STATS_SAMPLE_PAGES

STATUS

STDDEV_POP

STDDEV_SAMP

STDIN

STDOUT

STEP

STORAGE

STORAGE_INTEGRATION

STORAGE_SERIALIZATION_POLICY

STORED

STRAIGHT_JOIN

STREAM

STRICT

STRING

STRUCT

SUBMULTISET

SUBSTR

SUBSTRING

SUBSTRING_REGEX

SUCCEEDS

SUM

SUPER

SUPERUSER

SUPPORT

SUSPEND

SWAP

SYMMETRIC

SYNC

SYSTEM

SYSTEM_TIME

SYSTEM_USER

TABLE

TABLES

TABLESAMPLE

TABLESPACE

TAG

TARGET

TARGET_LAG

TASK

TBLPROPERTIES

TEMP

TEMPORARY

TEMPTABLE

TERMINATED

TERSE

TEXT

TEXTFILE

THEN

TIES

TIME

TIMEFORMAT

TIMESTAMP

TIMESTAMPTZ

TIMESTAMP_NTZ

TIMETZ

TIMEZONE

TIMEZONE_ABBR

TIMEZONE_HOUR

TIMEZONE_MINUTE

TIMEZONE_REGION

TINYBLOB

TINYINT

TINYTEXT

TO

TOP

TOTALS

TRACE

TRAILING

TRANSACTION

TRANSIENT

TRANSLATE

TRANSLATE_REGEX

TRANSLATION

TREAT

TRIGGER

TRIM

TRIM_ARRAY

TRUE

TRUNCATE

TRUNCATECOLUMNS

TRY

TRY_CAST

TRY_CONVERT

TSQUERY

TSVECTOR

TUPLE

TYPE

UBIGINT

UESCAPE

UHUGEINT

UINT8

UINT16

UINT32

UINT64

UINT128

UINT256

UNBOUNDED

UNCACHE

UNCOMMITTED

UNDEFINED

UNFREEZE

UNION

UNIQUE

UNKNOWN

UNLISTEN

UNLOAD

UNLOCK

UNLOGGED

UNMATCHED

UNNEST

UNPIVOT

UNSAFE

UNSET

UNSIGNED

UNTIL

UPDATE

UPPER

URL

USAGE

USE

USER

USER_RESOURCES

USING

USMALLINT

UTINYINT

UUID

VACUUM

VALID

VALIDATE

VALIDATION_MODE

VALUE

VALUES

VALUE_OF

VARBINARY

VARBIT

VARCHAR

VARIABLES

VARYING

VAR_POP

VAR_SAMP

VERBOSE

VERSION

VERSIONING

VERSIONS

VIEW

VIEWS

VIRTUAL

VOLATILE

VOLUME

WAREHOUSE

WAREHOUSES

WEEK

WEEKS

WHEN

WHENEVER

WHERE

WHILE

WIDTH_BUCKET

WINDOW

WITH

WITHIN

WITHOUT

WITHOUT_ARRAY_WRAPPER

WORK

WRAPPER

WRITE

XML

XMLNAMESPACES

XMLTABLE

XOR

YEAR

YEARS

ZONE

ZORDER

ZSTD