Module sqlparser::keywords

This module defines

a list of constants for every keyword

an ALL_KEYWORDS array with every keyword in it This is not a list of reserved keywords: some of these can be parsed as identifiers if the parser decides so. This means that new keywords can be added here without affecting the parse result.

As a matter of fact, most of these keywords are not used at all and could be removed.

a RESERVED_FOR_TABLE_ALIAS array with keywords reserved in a “table alias” context.

Enums

Keyword

Constants

ABORT

ABS

ABSOLUTE

ACTION

ADD

ADMIN

AGAINST

ALL

ALLOCATE

ALL_KEYWORDS

ALL_KEYWORDS_INDEX

ALTER

ANALYZE

AND

ANTI

ANY

APPLY

ARCHIVE

ARE

ARRAY

ARRAY_AGG

ARRAY_MAX_CARDINALITY

AS

ASC

ASENSITIVE

ASSERT

ASYMMETRIC

AT

ATOMIC

AUTHORIZATION

AUTOINCREMENT

AUTO_INCREMENT

AVG

AVRO

BACKWARD

BEGIN

BEGIN_FRAME

BEGIN_PARTITION

BETWEEN

BIGDECIMAL

BIGINT

BIGNUMERIC

BINARY

BLOB

BOOLEAN

BOTH

BTREE

BY

BYPASSRLS

BYTEA

CACHE

CALL

CALLED

CARDINALITY

CASCADE

CASCADED

CASE

CAST

CEIL

CEILING

CENTURY

CHAIN

CHANGE

CHAR

CHARACTER

CHARACTERS

CHARACTER_LENGTH

CHARSET

CHAR_LENGTH

CHECK

CLOB

CLONE

CLOSE

CLUSTER

COALESCE

COLLATE

COLLATION

COLLECT

COLUMN

COLUMNS

COMMENT

COMMIT

COMMITTED

COMPRESSION

COMPUTE

CONDITION

CONFLICT

CONNECT

CONNECTION

CONSTRAINT

CONTAINS

CONVERT

COPY

CORR

CORRESPONDING

COUNT

COVAR_POP

COVAR_SAMP

CREATE

CREATEDB

CREATEROLE

CROSS

CSV

CUBE

CUME_DIST

CURRENT

CURRENT_CATALOG

CURRENT_DATE

CURRENT_DEFAULT_TRANSFORM_GROUP

CURRENT_PATH

CURRENT_ROLE

CURRENT_ROW

CURRENT_SCHEMA

CURRENT_TIME

CURRENT_TIMESTAMP

CURRENT_TRANSFORM_GROUP_FOR_TYPE

CURRENT_USER

CURSOR

CYCLE

DATA

DATABASE

DATE

DATETIME

DAY

DEALLOCATE

DEC

DECADE

DECIMAL

DECLARE

DEFAULT

DELETE

DELIMITED

DELIMITER

DENSE_RANK

DEREF

DESC

DESCRIBE

DETERMINISTIC

DIRECTORY

DISCARD

DISCONNECT

DISTINCT

DISTRIBUTE

DO

DOUBLE

DOW

DOY

DROP

DUPLICATE

DYNAMIC

EACH

ELEMENT

ELSE

ENCODING

END

END_EXEC

END_FRAME

END_PARTITION

ENGINE

ENUM

EPOCH

EQUALS

ERROR

ESCAPE

EVENT

EVERY

EXCEPT

EXCLUDE

EXEC

EXECUTE

EXISTS

EXP

EXPANSION

EXPLAIN

EXTENDED

EXTERNAL

EXTRACT

FAIL

FALSE

FETCH

FIELDS

FILE

FILTER

FIRST

FIRST_VALUE

FLOAT

FLOOR

FOLLOWING

FOR

FORCE

FORCE_NOT_NULL

FORCE_NULL

FORCE_QUOTE

FOREIGN

FORMAT

FORWARD

FRAME_ROW

FREE

FREEZE

FROM

FULL

FULLTEXT

FUNCTION

FUNCTIONS

FUSION

GET

GLOBAL

GRANT

GRANTED

GRAPHVIZ

GROUP

GROUPING

GROUPS

HASH

HAVING

HEADER

HIVEVAR

HOLD

HOUR

IDENTITY

IF

IGNORE

ILIKE

IMMUTABLE

IN

INCREMENT

INDEX

INDICATOR

INHERIT

INNER

INOUT

INPUTFORMAT

INSENSITIVE

INSERT

INT

INTEGER

INTERSECT

INTERSECTION

INTERVAL

INTO

IS

ISODOW

ISOLATION

ISOYEAR

JAR

JOIN

JSON

JSONFILE

JULIAN

KEY

KILL

LAG

LANGUAGE

LARGE

LAST

LAST_VALUE

LATERAL

LEAD

LEADING

LEFT

LEVEL

LIKE

LIKE_REGEX

LIMIT

LISTAGG

LN

LOCAL

LOCALTIME

LOCALTIMESTAMP

LOCATION

LOCKED

LOGIN

LOWER

MANAGEDLOCATION

MATCH

MATCHED

MATERIALIZED

MAX

MAXVALUE

MEDIUMINT

MEMBER

MERGE

METADATA

METHOD

MICROSECOND

MICROSECONDS

MILLENIUM

MILLENNIUM

MILLISECOND

MILLISECONDS

MIN

MINUTE

MINVALUE

MOD

MODE

MODIFIES

MODULE

MONTH

MSCK

MULTISET

MUTATION

NANOSECOND

NANOSECONDS

NATIONAL

NATURAL

NCHAR

NCLOB

NEW

NEXT

NO

NOBYPASSRLS

NOCREATEDB

NOCREATEROLE

NOINHERIT

NOLOGIN

NONE

NOREPLICATION

NORMALIZE

NOSCAN

NOSUPERUSER

NOT

NOTHING

NOWAIT

NTH_VALUE

NTILE

NULL

NULLIF

NULLS

NUMERIC

NVARCHAR

OBJECT

OCCURRENCES_REGEX

OCTETS

OCTET_LENGTH

OF

OFFSET

OLD

ON

ONLY

OPEN

OPERATOR

OPTION

OPTIONS

OR

ORC

ORDER

OUT

OUTER

OUTPUTFORMAT

OVER

OVERFLOW

OVERLAPS

OVERLAY

OVERWRITE

OWNED

PARAMETER

PARQUET

PARTITION

PARTITIONED

PARTITIONS

PASSWORD

PERCENT

PERCENTILE_CONT

PERCENTILE_DISC

PERCENT_RANK

PERIOD

PLACING

PLANS

PORTION

POSITION

POSITION_REGEX

POWER

PRECEDES

PRECEDING

PRECISION

PREPARE

PRESERVE

PRIMARY

PRIOR

PRIVILEGES

PROCEDURE

PROGRAM

PURGE

QUALIFY

QUARTER

QUERY

QUOTE

RANGE

RANK

RCFILE

READ

READS

REAL

RECURSIVE

REF

REFERENCES

REFERENCING

REGCLASS

REGR_AVGX

REGR_AVGY

REGR_COUNT

REGR_INTERCEPT

REGR_R2

REGR_SLOPE

REGR_SXX

REGR_SXY

REGR_SYY

RELATIVE

RELEASE

RENAME

REPAIR

REPEATABLE

REPLACE

REPLICATION

RESERVED_FOR_COLUMN_ALIAS

Can’t be used as a column alias, so that SELECT <expr> alias can be parsed unambiguously without looking ahead.

RESERVED_FOR_TABLE_ALIAS

These keywords can’t be used as a table alias, so that FROM table_name alias can be parsed unambiguously without looking ahead.

RESTRICT

RESULT

RETURN

RETURNING

RETURNS

REVOKE

RIGHT

ROLE

ROLLBACK

ROLLUP

ROW

ROWID

ROWS

ROW_NUMBER

SAFE_CAST

SAVEPOINT

SCHEMA

SCOPE

SCROLL

SEARCH

SECOND

SELECT

SEMI

SENSITIVE

SEQUENCE

SEQUENCEFILE

SEQUENCES

SERDE

SERIALIZABLE

SESSION

SESSION_USER

SET

SETS

SHARE

SHOW

SIMILAR

SKIP

SMALLINT

SNAPSHOT

SOME

SORT

SPATIAL

SPECIFIC

SPECIFICTYPE

SQL

SQLEXCEPTION

SQLSTATE

SQLWARNING

SQRT

STABLE

START

STATIC

STATISTICS

STDDEV_POP

STDDEV_SAMP

STDIN

STDOUT

STORED

STRING

SUBMULTISET

SUBSTRING

SUBSTRING_REGEX

SUCCEEDS

SUM

SUPER

SUPERUSER

SYMMETRIC

SYNC

SYSTEM

SYSTEM_TIME

SYSTEM_USER

TABLE

TABLES

TABLESAMPLE

TBLPROPERTIES

TEMP

TEMPORARY

TEXT

TEXTFILE

THEN

TIES

TIME

TIMESTAMP

TIMESTAMPTZ

TIMETZ

TIMEZONE

TIMEZONE_HOUR

TIMEZONE_MINUTE

TINYINT

TO

TOP

TRAILING

TRANSACTION

TRANSIENT

TRANSLATE

TRANSLATE_REGEX

TRANSLATION

TREAT

TRIGGER

TRIM

TRIM_ARRAY

TRUE

TRUNCATE

TRY_CAST

TYPE

UESCAPE

UNBOUNDED

UNCACHE

UNCOMMITTED

UNION

UNIQUE

UNKNOWN

UNLOGGED

UNNEST

UNSIGNED

UNTIL

UPDATE

UPPER

USAGE

USE

USER

USING

UUID

VALID

VALUE

VALUES

VALUE_OF

VARBINARY

VARCHAR

VARIABLES

VARYING

VAR_POP

VAR_SAMP

VERBOSE

VERSIONING

VIEW

VIRTUAL

VOLATILE

WEEK

WHEN

WHENEVER

WHERE

WIDTH_BUCKET

WINDOW

WITH

WITHIN

WITHOUT

WORK

WRITE

XOR

YEAR

ZONE